### 2019-12-12 #1949 @whedon
[PRE REVIEW]: SampleDB: A sample and measurement metadata database
**Submitting author:** @FlorianRhiem (<a href="http://orcid.org/0000-0001-6461-9433">Florian Rhiem</a>)
**Repository:** <a href="https://github.com/sciapp/sampledb" target ="_blank">https://github.com/sciapp/sampledb</a>
**Version:** v0.8.1
