### 2019-12-14 #1949 @trallard
Hi @FlorianRhiem 👋  I will be serving as the editor for this submission.
The first thing we have to do is find 2 or more reviewers for this.

Do you already have some folks in mind? if not you can check our list of reviewers available here 👉🏼 <https://bit.ly/joss-reviewers>
