+++ sampledb/models/users.py
@@ -22,6 +22,7 @@ class User(db.Model, flask_login.UserMixin):
     email = db.Column(db.String, nullable=False)
     type = db.Column(db.Enum(UserType), nullable=False)
     is_admin = db.Column(db.Boolean, default=False, nullable=False)
+    is_readonly = db.Column(db.Boolean, default=False, nullable=False)
 
     def __init__(self, name, email, type):
         self.name = name
@@ -34,7 +35,9 @@ class User(db.Model, flask_login.UserMixin):
             self.name == other.name and
             self.email == other.email and
             self.type == other.type and
-            self.is_admin == other.is_admin)
+            self.is_admin == other.is_admin and
+            self.is_readonly == other.is_readonly
+        )
 
     def get_id(self):
         return self.id
