+++ sampledb/frontend/users/favorites.py
@@ -9,11 +9,13 @@ import flask_login
 from .. import frontend
 from .forms import ToggleFavoriteActionForm, ToggleFavoriteInstrumentForm
 from ...logic.favorites import add_favorite_action, remove_favorite_action, get_user_favorite_action_ids, add_favorite_instrument, remove_favorite_instrument, get_user_favorite_instrument_ids
+from ..utils import check_current_user_is_not_readonly
 
 
 @frontend.route('/users/me/favorite_actions/', methods=['POST'])
 @flask_login.login_required
 def toggle_favorite_action():
+    check_current_user_is_not_readonly()
     toggle_favorite_action_form = ToggleFavoriteActionForm()
     if toggle_favorite_action_form.validate():
         action_id = toggle_favorite_action_form.action_id.data
@@ -36,6 +38,7 @@ def toggle_favorite_action():
 @frontend.route('/users/me/favorite_instruments/', methods=['POST'])
 @flask_login.login_required
 def toggle_favorite_instrument():
+    check_current_user_is_not_readonly()
     toggle_favorite_instrument_form = ToggleFavoriteInstrumentForm()
     if toggle_favorite_instrument_form.validate():
         instrument_id = toggle_favorite_instrument_form.instrument_id.data
