### 2019-12-18 f46afca
Switch from GitHub packages to Docker Hub
