### 2019-12-20 #1949 @trallard
Ok let's ping the reviewers now :sparkles:

@eteq Erik, would you be able and willing to review this submission?
@usethedata Bruce, would you be able and willing to review this submission?
@jsta Joseph, would you be able and willing to review this submission?

Some important information:
- Our reviewer guidelines can be found here <https://joss.readthedocs.io/en/latest/review_criteria.html>
- **Title**:  SampleDB: A sample and measurement metadata database
- **GitHub repository**: https://github.com/sciapp/sampledb
-  **Submitting author**: @FlorianRhiem Florian Rhiem

We are looking at finding two reviewers to move this submission forward.
