+++ sampledb/logic/users.py
@@ -19,12 +19,15 @@ def get_user(user_id: int) -> User:
     return user
 
 
-def get_users() -> typing.List[User]:
+def get_users(exclude_hidden: bool = False) -> typing.List[User]:
     """
     Returns all users.
 
+    :param exclude_hidden: whether or not to exclude hidden users
     :return: the list of users
     """
+    if exclude_hidden:
+        return User.query.filter_by(is_hidden=False).all()
     return User.query.all()
 
 
@@ -69,5 +72,21 @@ def set_user_readonly(user_id: int, readonly: bool) -> None:
 
     user = get_user(user_id)
     user.is_readonly = readonly
+    db.session.add(user)
+    db.session.commit()
+
+
+def set_user_hidden(user_id: int, hidden: bool) -> None:
+    """
+    Set whether a user should be hidden from user lists.
+
+    :param user_id: the user ID of an existing user
+    :param hidden: True, if the user should be hidden, False otherwise
+    :raise errors.UserDoesNotExistError: when no user with the given
+        user ID exists
+    """
+
+    user = get_user(user_id)
+    user.is_hidden = hidden
     db.session.add(user)
     db.session.commit()
