+++ docs/user_guide/users.rst
@@ -73,3 +73,8 @@ Readonly Users
 --------------
 
 Users can be limited to READ permissions, e.g. for former employees who should still have access to their data but should not be able to create new SampleDB entries.
+
+Hidden users
+------------
+
+Users can also be hidden from users lists, which may be useful in similar use cases as when marking a user as readonly. These users can still be seen as part of an object's history or as members of groups and projects, but they will not be shown in the central users list, when granting permissions, inviting a user to a group or project, etc.
