+++ sampledb/frontend/users/preferences.py
@@ -3,6 +3,8 @@
 
 """
 
+import secrets
+
 import flask
 import flask_login
 import sqlalchemy.sql.expression
@@ -13,10 +15,10 @@ from .. import frontend
 from ..authentication_forms import ChangeUserForm, AuthenticationForm, AuthenticationMethodForm
 from ..users_forms import RequestPasswordResetForm, PasswordForm, AuthenticationPasswordForm
 from ..objects_forms import ObjectPermissionsForm, ObjectUserPermissionsForm, ObjectGroupPermissionsForm, ObjectProjectPermissionsForm
-from .forms import NotificationModeForm, OtherSettingsForm
+from .forms import NotificationModeForm, OtherSettingsForm, CreateAPITokenForm
 
 from ...logic import user_log
-from ...logic.authentication import add_authentication_method, remove_authentication_method, change_password_in_authentication_method
+from ...logic.authentication import add_authentication_method, remove_authentication_method, change_password_in_authentication_method, add_api_token
 from ...logic.users import get_user, get_users
 from ...logic.utils import send_confirm_email, send_recovery_email
 from ...logic.security_tokens import verify_token
@@ -59,13 +61,17 @@ def user_preferences(user_id):
 
 
 def change_preferences(user, user_id):
-    authentication_methods = Authentication.query.filter(Authentication.user_id == user_id).all()
+    api_tokens = Authentication.query.filter(Authentication.user_id == user_id, Authentication.type == AuthenticationType.API_TOKEN).all()
+    authentication_methods = Authentication.query.filter(Authentication.user_id == user_id, Authentication.type != AuthenticationType.API_TOKEN).all()
     authentication_method_ids = [authentication_method.id for authentication_method in authentication_methods]
-    confirmed_authentication_methods = Authentication.query.filter(Authentication.user_id == user_id, Authentication.confirmed == sqlalchemy.sql.expression.true()).count()
+    confirmed_authentication_methods = Authentication.query.filter(Authentication.user_id == user_id, Authentication.confirmed == sqlalchemy.sql.expression.true(), Authentication.type != AuthenticationType.API_TOKEN).count()
     change_user_form = ChangeUserForm()
     authentication_form = AuthenticationForm()
     authentication_method_form = AuthenticationMethodForm()
     authentication_password_form = AuthenticationPasswordForm()
+
+    created_api_token = None
+    create_api_token_form = CreateAPITokenForm()
 
     add_user_permissions_form = ObjectUserPermissionsForm()
     add_group_permissions_form = ObjectGroupPermissionsForm()
@@ -139,8 +145,9 @@ def change_preferences(user, user_id):
                                              public_permissions=public_permissions,
                                              authentication_method_form=authentication_method_form,
                                              authentication_form=authentication_form,
+                                             create_api_token_form=create_api_token_form,
                                              confirmed_authentication_methods=confirmed_authentication_methods,
-                                             authentications=authentication_methods, error=str(e))
+                                             authentications=authentication_methods, error=str(e), api_tokens=api_tokens)
             user_log.edit_user_preferences(user_id=user_id)
             return flask.redirect(flask.url_for('frontend.user_me_preferences'))
         else:
@@ -188,8 +195,9 @@ def change_preferences(user, user_id):
                                              public_permissions=public_permissions,
                                              authentication_method_form=authentication_method_form,
                                              authentication_form=authentication_form,
+                                             create_api_token_form=create_api_token_form,
                                              confirmed_authentication_methods=confirmed_authentication_methods,
-                                             authentications=authentication_methods, error=str(e))
+                                             authentications=authentication_methods, error=str(e), api_tokens=api_tokens)
[…]
