### 2020-01-15 ae2245b
Require sign in refresh for user preferences
