### 2020-01-15 0591756
Improve demo autologin
