### 2020-01-16 085f146
Use SERVER_PATH when determining whether a URL is safe
