### 2020-02-03 f8d69f7
Allow referencing measurements as object properties
