### 2020-02-03 #1949 @trallard
@celliern, @dvanic, @remram44 @fabioanomenegidio, @astrisovic I am the editor for this submission and we are looking for at least 2 reviewers. Would you be able/willing to review this submission?

Some important information:

Our reviewer guidelines can be found here <https://joss.readthedocs.io/en/latest/review_criteria.html>

- Title: SampleDB: A sample and measurement metadata database
- GitHub repository: <https://github.com/sciapp/sampledb>
- Submitting author: @FlorianRhiem Florian Rhiem
