### 2020-02-07 706e0c9
Do not suggest for objects to reference themselves
