### 2020-02-15 #2107 @whedon
[REVIEW]: SampleDB: A sample and measurement metadata database
**Submitting author:** @FlorianRhiem (<a href="http://orcid.org/0000-0001-6461-9433">Florian Rhiem</a>)
**Repository:** <a href="https://github.com/sciapp/sampledb" target ="_blank">https://github.com/sciapp/sampledb</a>
**Version:** 0.17.0
