### 2020-03-03 2bf4c92
Allow creating and editing instruments using the web frontend
