### 2020-03-03 ddea4f7
Set POSTGRES_HOST_AUTH_METHOD for GitLab CI
