### 2020-03-04 79568a1
Allow administrators to create actions without a user prefix
