### 2020-03-05 4eac92e
Allow administrators to mark users as hidden or read only using the web frontend
