### 2020-03-06 119970d
Added changelog entry for administration functions
