### 2020-03-09 2ad21a2
Enforce read only status for action and instrument creation and editing
