### 2020-03-11 #2107 @FlorianRhiem
@trallard SampleDB v0.9.0 has been released yesterday. I've updated the [`paper.json`](https://github.com/sciapp/sampledb/blob/develop/docs/publications/joss/paper.json). Could you please run `@whedon set v0.9.0 as version` to update the version here?
