### 2020-03-24 592f663
Added SAMPLEDB_DEMO_REDIRECT_URI
