### 2020-03-24 ed10b45
Allow configuring label formats
