+++ sampledb/frontend/objects.py
@@ -7,6 +7,7 @@ from copy import deepcopy
 import datetime
 import io
 import json
+import math
 import os
 import flask
 import flask_login
@@ -35,7 +36,7 @@ from .objects_forms import ObjectPermissionsForm, ObjectForm, ObjectVersionResto
 from ..utils import object_permissions_required
 from .utils import jinja_filter, generate_qrcode
 from .object_form_parser import parse_form_data
-from .labels import create_labels
+from .labels import create_labels, PAGE_SIZES, DEFAULT_PAPER_FORMAT, HORIZONTAL_LABEL_MARGIN, VERTICAL_LABEL_MARGIN, mm
 from .pdfexport import create_pdfexport
 from .utils import check_current_user_is_not_readonly
 
@@ -840,6 +841,10 @@ def object(object_id):
             get_object_location_assignment=get_object_location_assignment,
             get_user=get_user,
             get_location=get_location,
+            PAGE_SIZES=PAGE_SIZES,
+            HORIZONTAL_LABEL_MARGIN=HORIZONTAL_LABEL_MARGIN,
+            VERTICAL_LABEL_MARGIN=VERTICAL_LABEL_MARGIN,
+            mm=mm,
             object_location_assignments=get_object_location_assignments(object_id),
             build_object_location_assignment_confirmation_url=build_object_location_assignment_confirmation_url,
             user_may_assign_location=user_may_edit,
@@ -856,6 +861,79 @@ def object(object_id):
 @frontend.route('/objects/<int:object_id>/label')
 @object_permissions_required(Permissions.READ, on_unauthorized=on_unauthorized)
 def print_object_label(object_id):
+    mode = flask.request.args.get('mode', 'mixed')
+    if mode == 'fixed-width':
+        create_mixed_labels = False
+        create_long_labels = False
+        include_qrcode_in_long_labels = None
+        paper_format = flask.request.args.get('width-paper-format', '')
+        if paper_format not in PAGE_SIZES:
+            paper_format = DEFAULT_PAPER_FORMAT
+        maximum_width = math.floor(PAGE_SIZES[paper_format][0] / mm - 2 * HORIZONTAL_LABEL_MARGIN)
+        maximum_height = math.floor(PAGE_SIZES[paper_format][1] / mm - 2 * VERTICAL_LABEL_MARGIN)
+        ghs_classes_side_by_side = 'side-by-side' in flask.request.args
+        label_minimum_width = 20
+        if ghs_classes_side_by_side:
+            label_minimum_width = 40
+        try:
+            label_width = float(flask.request.args.get('label-width', '20'))
+        except ValueError:
+            label_width = 0
+        if math.isnan(label_width):
+            label_width = 0
+        if label_width < label_minimum_width:
+            label_width = label_minimum_width
+        if label_width > maximum_width:
+            label_width = maximum_width
+        try:
+            label_minimum_height = float(flask.request.args.get('label-minimum-height', '0'))
+        except ValueError:
+            label_minimum_height = 0
+        if math.isnan(label_minimum_height):
+            label_minimum_height = 0
+        if label_minimum_height < 0:
+            label_minimum_height = 0
+        if label_minimum_height > maximum_height:
+            label_minimum_height = maximum_height
+        qrcode_width = 18
+        centered = 'centered' in flask.request.args
+    elif mode == 'minimum-height':
+        create_mixed_labels = False
+        create_long_labels = True
+        paper_format = flask.request.args.get('height-paper-format', '')
+        if paper_format not in PAGE_SIZES:
+            paper_format = DEFAULT_PAPER_FORMAT
+        maximum_width = math.floor(PAGE_SIZES[paper_format][0] / mm - 2 * HORIZONTAL_LABEL_MARGIN)
[…]
