+++ sampledb/api/server/objects.py
@@ -10,6 +10,7 @@ from flask_restful import Resource
 
 from sampledb.api.server.authentication import multi_auth, object_permissions_required, Permissions
 from sampledb.logic.actions import get_action
+from sampledb.logic.object_search import generate_filter_func, wrap_filter_func
 from sampledb.logic.objects import get_object, update_object, create_object
 from sampledb.logic.object_permissions import get_objects_with_permissions
 from sampledb.logic import errors
@@ -117,13 +118,25 @@ class Object(Resource):
 class Objects(Resource):
     @multi_auth.login_required
     def get(self):
-        # TODO: implement filters
-        def filter_func(data):
-            return True
         action_id = None
         action_type = None
         project_id = None
-        search_notes = []
+        query_string = flask.request.args.get('q', '')
+        if query_string:
+            try:
+                filter_func, search_tree, use_advanced_search = generate_filter_func(query_string, True)
+            except Exception:
+                # TODO: ensure that advanced search does not cause exceptions
+                def filter_func(data, search_notes):
+                    """ Return all objects"""
+                    search_notes.append(('error', "Unable to parse search expression", 0, len(query_string)))
+                    return False
+            filter_func, search_notes = wrap_filter_func(filter_func)
+        else:
+            search_notes = []
+
+            def filter_func(data):
+                return True
         try:
             objects = get_objects_with_permissions(
                 user_id=flask.g.user.id,
@@ -136,17 +149,21 @@ class Objects(Resource):
         except Exception as e:
             search_notes.append(('error', "Error during search: {}".format(e), 0, 0))
             objects = []
-        # TODO handle search notes and set error code
-        return [
-            {
-                'object_id': object.object_id,
-                'version_id': object.version_id,
-                'action_id': object.action_id,
-                'schema': object.schema,
-                'data': object.data
-            }
-            for object in objects
-        ]
+        if any(search_note[0] == 'error' for search_note in search_notes):
+            return [
+                [search_note[0], search_note[1]] for search_note in search_notes
+            ], 400
+        else:
+            return [
+                {
+                    'object_id': object.object_id,
+                    'version_id': object.version_id,
+                    'action_id': object.action_id,
+                    'schema': object.schema,
+                    'data': object.data
+                }
+                for object in objects
+            ], 200
 
     @multi_auth.login_required
     def post(self):
