### 2020-04-17 9b142a7
Document search filters for API
