### 2020-04-17 21e77ec
Fix invalid format usage caught by pyflakes
