### 2020-05-04 07f03e4
Fix API error messages for object search
