### 2020-05-04 6a7bb02
Improved error messages for API object search
