### 2020-05-11 #2107 @FlorianRhiem
@trallard I have just released SampleDB v0.10.0. Could you please run `@whedon set v0.10.0 as version` so that it is updated here? Thank you!
