### 2020-05-13 ed3d76c
Add password to postgres example in README
