+++ sampledb/version.py
@@ -1,0 +1,8 @@
+# coding: utf-8
+"""
+This file should only contain the definition of the __version__ variable.
+
+It may be read by multiple other files, e.g. setup.py and the status page.
+"""
+
+__version__ = '0.10.0'
