### 2020-05-25 cac051a
Only recommend dev-sampledb if docs are built for iffSamples
