+++ .gitlab-ci.yml
@@ -57,7 +57,7 @@ documentation:
     - export DISPLAY=:0
     - python3 docs/utils/generate_images.py
     # build documentation
-    - python3 -m sphinx -b html docs/ build_documentation/
+    - python3 -m sphinx -b html -t iffSamples docs/ build_documentation/
   artifacts:
     paths:
     - build_documentation
