### 2020-06-08 c280e5f
Fix typos in API documentation
