### 2020-06-10 #2107 @dvanic
# Review checklist for @dvanic
## Conflict of interest

- [x] I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

- [x] I confirm that I read and will adhere to the JOSS code of conduct.

- [x] Repository: Is the source code for this software available at the repository url?
- [x] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

- [x] Contribution and authorship: Has the submitting author (@FlorianRhiem) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

     - Note: not yet. There are challenges with documentation around setting up new users, groups and projects.

This is the area I struggled with the most: I could not identify clear documentation on how to set up a completely new instance of sampleDB at a new institution, add users, groups, projects, metadata, samples, instruments etc, and then how to proceed to a tracked lifecycle of samples and visualisations in Jupyter.

Please note I followed the "Getting Started" instructions in the GitHub repository, and while I could try to read the output options for all of the possible commands in `docker exec sampledb env/bin/python -m sampledb help` it's not clear what the optimal/recommended order of setup is. To enable wide adoption of this tool a case study of setting up a basic database is vital (because I genuinely think this is a very useful software, that really could be adopted beyond the institution in which it was developed!).

API documentation is available. Due to not being able to set up or access a dummy database I was unable to test it.

- [x] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

It is not clear how users are meant to seek support: also via GitHub issues?

- [x] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

Yes, although the database could be of use beyond physics (for ex. in the life sciences, especially biology), and I'm not sure the provided description would enable life science users to see this.

Yes.

- [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?

No. No mention of other packages is made.

- [x] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

- [ ] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

No scientific publications are cited - only references to software tools that underpin the database, and a metapaper around research data management.

## Footnote

I'd like to personally apologize to @FlorianRhiem for the amount of time it's taken me to review this. While I'm fortunate enough to be in Australia, where COVID has not raged as widely as in many other parts of the world, our workload this semester has been horrific due to the move to online teaching/training (as well as the challenges of childcare and life). @FlorianRhiem I am SO sorry for how long this review has taken! And while I have made some suggestions, I think this is a very useful piece of software, and one we can potentially adapt to some of the work we do with researchers at the Sydney Informatics Hub/University of Sydney.
