### 2020-06-10 #2107 @FlorianRhiem
Thank you for the review, Darya. I'm glad to hear that you think SampleDB could be useful for more researchers and I will try to fix these issues until the end of next week.

I think the most difficult points will be the documentation for more than the minimal setup and for the JupyterHub integration. For both there are dependencies to likely existing IT infrastructure (where should backups be stored, which volumes are mounted by the JupyterHub notebook servers, etc.), but I will try to keep the information as independent of these details as possible. Currently the online documentation is only updated for the tagged releases, so if that's alright with you, I will post drafts for these texts in the GitHub issues you opened, so that you can let me know whether something is missing or too vague.
