### 2020-06-10 dc3387d
Create admin user from config values during initial setup
