### 2020-06-15 095c4eb
Rename guest invitation if LDAP is not configured
