### 2020-06-16 5ac40cf
Added administrator guide
