+++ docs/user_guide/instruments.rst
@@ -10,6 +10,8 @@ You can view a list of instruments at |service_instruments_url|. To make navigat
 .. note::
     At this time, instruments can only be created by the |service_name| administrators. If you would like your instrument or action to be included, please `let us know`_.
 
+.. _instrument_scientists:
+
 Instrument Scientists
 ---------------------
 
