+++ docs/developer_guide/contributing.rst
@@ -1,0 +1,6 @@
+.. _contributing:
+
+Contributing
+============
+
+You can find information on how to report issues, set up a development installation or submit changes in the `Contribution Guide <https://github.com/sciapp/sampledb/blob/develop/CONTRIBUTING.md>`_ in the `SampleDB GitHub repository <https://github.com/sciapp/sampledb>`_.
