+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Currently in development.
 
 - Allow usage of Markdown in instrument and action descriptions
 - Added configuration values for creating an admin user during initial setup
+- Added administrator guide to documentation
 
 Version 0.10
 ------------

[…]
