### 2020-06-17 2220043
Added information on alternatives to SampleDB
