+++ docs/publications/joss/paper.bib
@@ -26,3 +26,24 @@
   url = {https://pypi.org/project/SQLAlchemy/},
   urldate = {2019-12-10}
 }
+@online{juliabase,
+  author = {Torsten Bronger},
+  title = {The JuliaBase project},
+  year = 2015,
+  url = {https://juliabase.org/project.html},
+  urldate = {2020-06-17}
+}
+@online{bikalims,
+  author = {{Bika LIMS Collective}},
+  title = {The Bika Open Source LIMS project},
+  year = 2020,
+  url = {https://www.bikalims.org/},
+  urldate = {2020-06-17}
+}
+@online{senaite,
+  author = {{RIDING BYTES GmbH} and {NARALABS S.L.}},
+  title = {SENAITE Enterprise Open Source Laboratory System},
+  year = 2020,
+  url = {https://www.senaite.com/},
+  urldate = {2020-06-17}
+}
