### 2020-06-17 9385725
Added fair principles instead of only accessibility
