+++ docs/publications/joss/paper.bib
@@ -47,3 +47,13 @@
   url = {https://www.senaite.com/},
   urldate = {2020-06-17}
 }
+@article{fair,
+  author  = {Mark D. Wilkinson and  Michel Dumontier and IJsbrand Jan Aalbersberg and Gabrielle Appleton and Myles Axton and Arie Baak and Niklas Blomberg and Jan-Willem Boiten and Luiz Bonino da Silva Santos and Philip E. Bourne and Jildau Bouwman and Anthony J. Brookes and Tim Clark and Mercè Crosas and Ingrid Dillo and Olivier Dumon and Scott Edmunds and Chris T. Evelo and Richard Finkers and Alejandra Gonzalez-Beltran and Alasdair J.G. Gray and Paul Groth and Carole Goble and Jeffrey S. Grethe and Jaap Heringa and Peter A.C ’t Hoen and Rob Hooft and Tobias Kuhn and Ruben Kok and Joost Kok and Scott J. Lusher and Maryann E. Martone and Albert Mons and Abel L. Packer and Bengt Persson and Philippe Rocca-Serra and Marco Roos and Rene van Schaik and Susanna-Assunta Sansone and Erik Schultes and Thierry Sengstag and Ted Slater and George Strawn and Morris A. Swertz and Mark Thompson and Johan van der Lei and Erik van Mulligen and Jan Velterop and Andra Waagmeester and Peter Wittenburg and Katherine Wolstencroft and Jun Zhao and Barend Mons},
+  title   = {The {FAIR Guiding Principles} for scientific data management and stewardship},
+  journal = {Scientific Data},
+  volume  = {3},
+  year    = {2016},
+  number  = {160018},
+  doi     = {10.1038/sdata.2016.18},
+  url     = {https://doi.org/10.1038/sdata.2016.18}
+}
