+++ docs/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 Version 0.11
 ------------
 
-Currently in development.
+Released on June 18th, 2020.
 
 - Allow usage of Markdown in instrument and action descriptions
 - Added configuration values for creating an admin user during initial setup
