### 2020-06-18 #2107 @FlorianRhiem
I have created an Administrator Guide with a new [Getting Started](https://scientific-it-systems.iffgit.fz-juelich.de/SampleDB/administrator_guide/getting_started.html) guide, which walks readers through the first steps. I hope this solves the functionality and documentation issues for setting up SampleDB. If information on a particular step is still missing or too vague, please let me know!

I have added a Support section to the README and also recommend opening GitHub issues at the end of the Getting Started guide.

I have updated the paper to include a paragraph in which I discuss what I see as the largest difference between SampleDB and other applications with similar goals, in particular JuliaBase and BikaLIMS / Senaite, and which advantages and disadvantages result from this. In terms of goals, JuliaBase is the closest to SampleDB, but BikaLIMS / Senaite include sample management as a core feature as well and I think it's worth mentioning them as their approach of having pre-defined calculations might be something worth exploring in the future, for use cases where a data analysis in a Jupyter notebook might not be necessary.

I have also added a reference to [The FAIR Guiding Principles for scientific data management and stewardship](https://www.nature.com/articles/sdata201618) and updated the text to include more than just making the data accessible, as the motivation and reasoning behind the FAIR principles as presented in the paper match the motivation behind SampleDB as well.

I hope I have addressed the issues with SampleDB and the paper, either here or in the GitHub Issues of the project. Please let me know what I can improve and thank you again for taking the time to review the paper, particular given the current situation.

---

@trallard I've released SampleDB 0.11.0, could you please updated the version with whedon? Thank you!
