### 2020-06-19 e23488f
Allow adding a logo to object export PDF documents
