### 2020-06-19 ee5d303
Added utility function for ANSI color codes
