### 2020-06-19 2f11580
Add check for FILE_STORAGE_PATH
