+++ tests/logic/test_export.py
@@ -1,0 +1,199 @@
+# coding: utf-8
+"""
+
+"""
+
+import io
+import json
+import tarfile
+import zipfile
+
+import pytest
+
+import sampledb
+from sampledb.models import User, ActionType
+from sampledb.logic import export, objects, actions, files
+
+from ..test_utils import app_context, app, flask_server
+
+
+@pytest.fixture
+def user(flask_server):
+    with flask_server.app.app_context():
+        user = sampledb.models.User(name="Basic User", email="example@fz-juelich.de", type=sampledb.models.UserType.PERSON)
+        sampledb.db.session.add(user)
+        sampledb.db.session.commit()
+        # force attribute refresh
+        assert user.id is not None
+    return user
+
+
+def set_up_state(user: User):
+    action = actions.create_action(
+        action_type=ActionType.SAMPLE_CREATION,
+        name='Example Action',
+        schema={
+            'title': 'Example Object',
+            'type': 'object',
+            'properties': {
+                'name': {
+                    'title': 'Sample Name',
+                    'type': 'text'
+                }
+            },
+            'required': ['name']
+        },
+        description='',
+        instrument_id=None
+    )
+    actions.create_action(
+        action_type=ActionType.SAMPLE_CREATION,
+        name='Irrelevant Action',
+        schema={
+            'title': 'Example Object',
+            'type': 'object',
+            'properties': {
+                'name': {
+                    'title': 'Sample Name',
+                    'type': 'text'
+                }
+            },
+            'required': ['name']
+        },
+        description='',
+        instrument_id=None
+    )
+    data = {'name': {'_type': 'text', 'text': 'Object'}}
+    object = objects.create_object(user_id=user.id, action_id=action.id, data=data)
+    def save_content(file): file.write("This is a test file.".encode('utf-8'))
+    files.create_local_file(object.id, user.id, "test.txt", save_content)
+    files.create_url_file(object.id, user.id, "https://example.com")
+
+
+def validate_data(data):
+    # remove datetimes before comparison
+    del data['objects'][0]['versions'][0]['utc_datetime']
+    del data['objects'][0]['files'][0]['utc_datetime']
+    del data['objects'][0]['files'][1]['utc_datetime']
+
+    assert data == {
+        'objects': [
+            {
+                'id': 1,
+                'action_id': 1,
+                'versions': [
+                    {
+                        'id': 0,
+                        'user_id': 1,
+                        'schema': {
+                                'type': 'object',
+                                'title': 'Example Object',
+                                'required': ['name'],
+                                'properties': {
+                                    'name': {
+                                        'type': 'text',
+                                        'title': 'Sample Name'
+                                    }
+                                }
+                        },
+                        'data': {
+                            'name': {
+                                'text': 'Object',
+                                '_type': 'text'
+                            }
+                        }
+                    }
+                ],
+                'comments': [],
+                'location_assignments': [],
+                'publications': [],
+                'files': [
+                    {
+                        'id': 0,
+                        'hidden': False,
+                        'title': 'test.txt',
+                        'description': None,
+                        'uploader_id': 1,
+                        'original_file_name': 'test.txt',
+                        'path': 'objects/1/files/0/test.txt'
+                    },
+                    {
+                        'id': 1,
+                        'hidden': False,
+                        'title': 'https://example.com',
+                        'description': None,
[…]
