+++ docs/user_guide/objects.rst
@@ -162,12 +162,16 @@ When an object is created, its creator, any associated instrument scientists and
 
     Default Permissions in the User Preferences
 
+.. _pdf_export:
+
 PDF Export
 ----------
 
 Users can export object information to a PDF file, e.g. for printing or offline usage. Note that the exported object information will not be fully complete, e.g. only files of some formats will be included in the PDF and only the current metadata version will be shown.
 
 Along with the current object, related objects can be exported along with it, e.g. a sample can be exported as PDF together with all measurements performed with it.
+
+Users can also create a PDF file for all objects which the user has **READ** permissions for (see :ref:`export`).
 
 Labels
 ------
