### 2020-06-25 2cd5663
Allow setting a publicly visible ORCID iD
