+++ docs/changelog.rst
@@ -6,8 +6,9 @@ Version 0.12
 
 Currently in development.
 
-- Add data export as PDF document, .zip or .tar.gz archive
+- Added data export as PDF document, .zip or .tar.gz archive
 - Allow adding a logo to object export PDF documents
+- Allow setting a publicly visible ORCID iD
 
 Version 0.11
 ------------
