+++ sampledb/logic/instrument_log_entries.py
@@ -1,0 +1,244 @@
+# coding: utf-8
+"""
+Logic module for instrument log entries
+
+Instrument scientists can add messages to a log for their instruments. They
+can also allow users to provide log entries, e.g. to report issues. These log
+entries are immutable and therefore this module only allows the creation and
+querying of comments.
+
+Files can be attached to these log entries, e.g. to supply the instrument scientists
+with control files or photographs of an experiment setup.
+"""
+
+import collections
+import datetime
+import typing
+
+from .. import db
+from . import errors, instruments, users, user_log, notifications
+from ..models import instrument_log_entries
+
+
+class InstrumentLogEntry(
+    collections.namedtuple(
+        'InstrumentLogEntry',
+        ['id', 'instrument_id', 'user_id', 'content', 'utc_datetime']
+    )
+):
+    """
+    This class provides an immutable wrapper around models.instrument_log_entries.InstrumentLogEntry.
+    """
+
+    def __new__(
+            cls,
+            id: int,
+            instrument_id: int,
+            user_id: int,
+            content: str,
+            utc_datetime: datetime.datetime
+    ):
+        self = super(InstrumentLogEntry, cls).__new__(
+            cls, id, instrument_id, user_id, content, utc_datetime
+        )
+        return self
+
+    @classmethod
+    def from_database(cls, instrument_log_entry: instrument_log_entries.InstrumentLogEntry) -> 'InstrumentLogEntry':
+        return InstrumentLogEntry(
+            id=instrument_log_entry.id,
+            instrument_id=instrument_log_entry.instrument_id,
+            user_id=instrument_log_entry.user_id,
+            content=instrument_log_entry.content,
+            utc_datetime=instrument_log_entry.utc_datetime
+        )
+
+    @property
+    def author(self):
+        return users.get_user(self.user_id)
+
+    @property
+    def file_attachments(self) -> typing.List['InstrumentLogFileAttachment']:
+        return get_instrument_log_file_attachments(self.id)
+
+
+class InstrumentLogFileAttachment(
+    collections.namedtuple(
+        'InstrumentLogFileAttachment',
+        ['id', 'log_entry_id', 'file_name', 'content']
+    )
+):
+    """
+    This class provides an immutable wrapper around models.instrument_log_entries.InstrumentLogFileAttachment.
+    """
+
+    def __new__(
+            cls,
+            id: int,
+            log_entry_id: int,
+            file_name: str,
+            content: bytes
+    ):
+        self = super(InstrumentLogFileAttachment, cls).__new__(
+            cls, id, log_entry_id, file_name, content
+        )
+        return self
+
+    @classmethod
+    def from_database(
+            cls,
+            instrument_log_file_attachment: instrument_log_entries.InstrumentLogFileAttachment
+    ) -> 'InstrumentLogFileAttachment':
+        return InstrumentLogFileAttachment(
+            id=instrument_log_file_attachment.id,
+            log_entry_id=instrument_log_file_attachment.log_entry_id,
+            file_name=instrument_log_file_attachment.file_name,
+            content=instrument_log_file_attachment.content
+        )
+
+
+def create_instrument_log_entry(
+        instrument_id: int,
+        user_id: int,
+        content: str
+) -> InstrumentLogEntry:
+    """
+    Creates a new instrument log entry and adds it to the user log.
+
+    :param instrument_id: the ID of an existing instrument
+    :param user_id: the ID of an existing user
+    :param content: the text content for the new comment
+    :raise errors.InstrumentDoesNotExistError: when no instrument with the
+        given instrument ID exists
[…]
