### 2020-07-02 bec27a9
Make PDF export sections optional
