### 2020-07-02 57242b3
Add instrument scientist notes
