### 2020-07-10 6cecb1a
Improve layout of instrument notes
