### 2020-07-16 8ca941a
Allow empty log entries as long as files are attached
