### 2020-07-16 07a7923
Add instrument log object attachments
