+++ sampledb/logic/errors.py
@@ -196,3 +196,7 @@ class InstrumentLogEntryDoesNotExistError(Exception):
 
 class InstrumentLogFileAttachmentDoesNotExistError(Exception):
     pass
+
+
+class InstrumentLogObjectAttachmentDoesNotExistError(Exception):
+    pass
