### 2020-07-16 35bf478
Add instrument log categories
