+++ sampledb/logic/instrument_log_entries.py
@@ -18,12 +18,13 @@ import typing
 from .. import db
 from . import errors, instruments, users, user_log, notifications, objects
 from ..models import instrument_log_entries
+from ..models.instrument_log_entries import InstrumentLogCategoryTheme
 
 
 class InstrumentLogEntry(
     collections.namedtuple(
         'InstrumentLogEntry',
-        ['id', 'instrument_id', 'user_id', 'content', 'utc_datetime']
+        ['id', 'instrument_id', 'user_id', 'content', 'utc_datetime', 'categories']
     )
 ):
     """
@@ -36,10 +37,11 @@ class InstrumentLogEntry(
             instrument_id: int,
             user_id: int,
             content: str,
-            utc_datetime: datetime.datetime
+            utc_datetime: datetime.datetime,
+            categories: typing.List['InstrumentLogCategory']
     ):
         self = super(InstrumentLogEntry, cls).__new__(
-            cls, id, instrument_id, user_id, content, utc_datetime
+            cls, id, instrument_id, user_id, content, utc_datetime, categories
         )
         return self
 
@@ -50,7 +52,11 @@ class InstrumentLogEntry(
             instrument_id=instrument_log_entry.instrument_id,
             user_id=instrument_log_entry.user_id,
             content=instrument_log_entry.content,
-            utc_datetime=instrument_log_entry.utc_datetime
+            utc_datetime=instrument_log_entry.utc_datetime,
+            categories=[
+                InstrumentLogCategory.from_database(category)
+                for category in sorted(instrument_log_entry.categories, key=lambda category: category.theme.value)
+            ]
         )
 
     @property
@@ -134,31 +140,158 @@ class InstrumentLogObjectAttachment(
         )
 
 
+class InstrumentLogCategory(
+    collections.namedtuple(
+        'InstrumentLogCategory',
+        ['id', 'instrument_id', 'title', 'theme']
+    )
+):
+    """
+    This class provides an immutable wrapper around models.instrument_log_entries.InstrumentLogCategory.
+    """
+
+    def __new__(
+            cls,
+            id: int,
+            instrument_id: int,
+            title: str,
+            theme: instrument_log_entries.InstrumentLogCategoryTheme
+    ):
+        self = super(InstrumentLogCategory, cls).__new__(
+            cls, id, instrument_id, title, theme
+        )
+        return self
+
+    @classmethod
+    def from_database(
+            cls,
+            category: instrument_log_entries.InstrumentLogCategory
+    ) -> 'InstrumentLogCategory':
+        return InstrumentLogCategory(
+            id=category.id,
+            instrument_id=category.instrument_id,
+            title=category.title,
+            theme=category.theme
+        )
+
+
+def get_instrument_log_categories(instrument_id: int) -> typing.List[InstrumentLogCategory]:
+    """
+    Return the instrument log categories for an object.
+
+    :param instrument_id: the ID of an existing instrument
+    :return: the instrument log categories
+    :raise errors.InstrumentDoesNotExistError: when no instrument with the
+        given instrument ID exists
+    """
+    categories = instrument_log_entries.InstrumentLogCategory.query.filter_by(instrument_id=instrument_id).all()
+    if not categories:
+        # ensure that the instrument exists
+        instruments.get_instrument(instrument_id=instrument_id)
+        return []
+    return [
+        InstrumentLogCategory.from_database(category)
+        for category in categories
+    ]
+
+
+def create_instrument_log_category(
+    instrument_id: int,
+    title: str,
+    theme: instrument_log_entries.InstrumentLogCategoryTheme
+) -> InstrumentLogCategory:
+    """
+    Create a new instrument log category for an instrument.
+
+    :param instrument_id: the ID of an existing instrument
+    :param title: the category title
+    :param theme: the category theme
[…]
