+++ sampledb/models/instrument_log_entries.py
@@ -2,15 +2,50 @@
 """
 Model for instrument log entries
 
-This should not to be confused with internal logging like the object log
+This should not to be confused with internal logging like the object log.
 """
 
+import enum
 import datetime
 import typing
 
 from .. import db
 from .instruments import Instrument
 from .objects import Objects
+
+
+instrument_log_entry_category_association_table = db.Table(
+    'instrument_log_entry_category_associations',
+    db.metadata,
+    db.Column('log_entry_id', db.Integer, db.ForeignKey('instrument_log_entries.id')),
+    db.Column('category_id', db.Integer, db.ForeignKey('instrument_log_categories.id', ondelete="CASCADE"))
+)
+
+
+@enum.unique
+class InstrumentLogCategoryTheme(enum.Enum):
+    GRAY = 0
+    BLUE = 1
+    GREEN = 2
+    YELLOW = 3
+    RED = 4
+
+
+class InstrumentLogCategory(db.Model):
+    __tablename__ = 'instrument_log_categories'
+
+    id = db.Column(db.Integer, primary_key=True)
+    instrument_id = db.Column(db.Integer, db.ForeignKey(Instrument.id), nullable=False)
+    title = db.Column(db.String, nullable=False)
+    theme = db.Column(db.Enum(InstrumentLogCategoryTheme), nullable=False)
+
+    def __init__(self, instrument_id: int, title: str, theme: InstrumentLogCategoryTheme):
+        self.instrument_id = instrument_id
+        self.title = title
+        self.theme = theme
+
+    def __repr__(self):
+        return '<{0}(id={1.id}, instrument_id={1.instrument_id}, title="{1.title}", theme={1.theme.name.lower()})>'.format(type(self).__name__, self)
 
 
 class InstrumentLogEntry(db.Model):
@@ -22,6 +57,7 @@ class InstrumentLogEntry(db.Model):
     content = db.Column(db.Text, nullable=False)
     utc_datetime = db.Column(db.DateTime, nullable=False)
     author = db.relationship('User')
+    categories = db.relationship('InstrumentLogCategory', secondary=instrument_log_entry_category_association_table)
 
     def __init__(
             self,
