### 2020-07-20 0dafe59
Allow creating an instrument log entry while creating an object
