### 2020-07-20 0e9f72c
Show form error when instrument log entry text is missing
