### 2020-07-27 e075ebb
Added smartphone upload to instrument log
