+++ sampledb/frontend/instruments.py
@@ -6,6 +6,7 @@
 import json
 import os
 
+import itsdangerous
 import flask
 import flask_login
 from flask_wtf import FlaskForm
@@ -16,13 +17,13 @@ from . import frontend
 from ..logic.instruments import get_instruments, get_instrument, create_instrument, update_instrument, set_instrument_responsible_users
 from ..logic.instrument_log_entries import get_instrument_log_entries, create_instrument_log_entry, get_instrument_log_file_attachment, create_instrument_log_file_attachment, create_instrument_log_object_attachment, get_instrument_log_object_attachments, get_instrument_log_categories, InstrumentLogCategoryTheme, create_instrument_log_category, update_instrument_log_category, delete_instrument_log_category
 from ..logic.actions import ActionType
-from ..logic.errors import InstrumentDoesNotExistError, InstrumentLogFileAttachmentDoesNotExistError, ObjectDoesNotExistError
+from ..logic.errors import InstrumentDoesNotExistError, InstrumentLogFileAttachmentDoesNotExistError, ObjectDoesNotExistError, UserDoesNotExistError
 from ..logic.favorites import get_user_favorite_instrument_ids
-from ..logic.users import get_users
+from ..logic.users import get_users, get_user
 from ..logic.objects import get_object
 from ..logic.object_permissions import Permissions, get_objects_with_permissions
 from .users.forms import ToggleFavoriteInstrumentForm
-from .utils import check_current_user_is_not_readonly, markdown_to_safe_html
+from .utils import check_current_user_is_not_readonly, markdown_to_safe_html, generate_qrcode
 
 __author__ = 'Florian Rhiem <f.rhiem@fz-juelich.de>'
 
@@ -79,6 +80,14 @@ def instrument(instrument_id):
     else:
         instrument_log_entries = None
         attached_object_names = {}
+    if is_instrument_responsible_user or instrument.users_can_create_log_entries:
+        serializer = itsdangerous.URLSafeTimedSerializer(flask.current_app.config['SECRET_KEY'], salt='instrument-log-mobile-upload')
+        token = serializer.dumps([flask_login.current_user.id, instrument_id])
+        mobile_upload_url = flask.url_for('.instrument_log_mobile_file_upload', instrument_id=instrument_id, token=token, _external=True)
+        mobile_upload_qrcode = generate_qrcode(mobile_upload_url, should_cache=False)
+    else:
+        mobile_upload_url = None
+        mobile_upload_qrcode = None
     instrument_log_entry_form = InstrumentLogEntryForm()
     instrument_log_entry_form.objects.choices = [
         (str(object.id), "{} (#{})".format(object.data.get('name', {}).get('text', 'Unnamed Object'), object.id))
@@ -142,6 +151,8 @@ def instrument(instrument_id):
         attached_object_names=attached_object_names,
         is_instrument_responsible_user=is_instrument_responsible_user,
         instrument_log_entry_form=instrument_log_entry_form,
+        mobile_upload_url=mobile_upload_url,
+        mobile_upload_qrcode=mobile_upload_qrcode,
         ActionType=ActionType
     )
 
@@ -247,7 +258,7 @@ def edit_instrument(instrument_id):
         return flask.abort(404)
     check_current_user_is_not_readonly()
     if not flask_login.current_user.is_admin and flask_login.current_user not in instrument.responsible_users:
-        return flask.abort(401)
+        return flask.abort(403)
     instrument_form = InstrumentForm()
     instrument_form.name.default = instrument.name
     instrument_form.description.default = instrument.description
@@ -387,3 +398,70 @@ def instrument_log_file_attachment(instrument_id, log_entry_id, file_attachment_
                 'Content-Type': mime_type
             }
         )
+
+
[…]
