+++ sampledb/logic/rdf.py
@@ -1,0 +1,172 @@
+# coding: utf-8
+"""
+Dublin Core metadata in RDF/XML format
+"""
+
+import typing
+
+import flask
+import jinja2
+
+from . import objects, object_log, users
+
+RDF_TEMPLATE = jinja2.Template("""<?xml version="1.0"?>
+<rdf:RDF
+xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
+xmlns:dcterms="http://purl.org/dc/terms/"
+xmlns:foaf="http://xmlns.com/foaf/0.1/"
+>
+  <rdf:Description xml:lang="en" rdf:about="{{ object_url }}">
+    <dcterms:identifier>{{ object_url }}</dcterms:identifier>
+    <dcterms:title>{{object_name}}</dcterms:title>
+    <dcterms:created>{{ created.isoformat() }}</dcterms:created>
+    <dcterms:modified>{{ modified.isoformat() }}</dcterms:modified>
+
+{% for user in creators %}
+    <dcterms:creator>
+      <foaf:Person rdf:about="{{ user['url'] }}">
+        <foaf:name>{{ user['name'] }}</foaf:name>
+      </foaf:Person>
+    </dcterms:creator>
+{% endfor %}
+
+{% for user in contributors %}
+    <dcterms:contributor>
+      <foaf:Person rdf:about="{{ user['url'] }}">
+        <foaf:name>{{ user['name'] }}</foaf:name>
+      </foaf:Person>
+    </dcterms:contributor>
+{% endfor %}
+
+{% for version_url in version_urls %}
+    <dcterms:hasVersion rdf:resource="{{ version_url }}" />
+{% endfor %}
+
+{% if current_object_url %}
+    <dcterms:isVersionOf rdf:resource="{{ current_object_url }}" />
+{% endif %}
+  </rdf:Description>
+</rdf:RDF>
+""")
+
+
+def generate_rdf(user_id: int, object_id: int, version_id: typing.Optional[int] = None) -> str:
+    """
+    Generate an XML file for an object.
+
+    The file will contain an RDF description for an object using Dublin Core metadata.
+
+    :param user_id: the ID of an existing user
+    :param object_id: the ID of the existing object
+    :param version_id: the ID of the object's existing version
+    :return:
+    :raise errors.UserDoesNotExistError: when no user with the given ID exists
+    :raise errors.ObjectDoesNotExistError: when no object with the given
+        object ID exists
+    :raise errors.ObjectVersionDoesNotExistError: when an object with the
+        given object ID exists, but does not have a version with the given
+        version ID
+    """
+
+    object = objects.get_object(object_id, version_id)
+
+    log_entries = object_log.get_object_log_entries(object.id, user_id)
+
+    creation_datetime = object.utc_datetime
+    modification_datetime = object.utc_datetime
+
+    creator_ids = []
+    contributor_ids = []
+    for entry in reversed(log_entries):
+        if entry.utc_datetime > modification_datetime:
+            modification_datetime = entry.utc_datetime
+        if entry.type in {
+            object_log.ObjectLogEntryType.CREATE_BATCH,
+            object_log.ObjectLogEntryType.CREATE_OBJECT,
+            object_log.ObjectLogEntryType.EDIT_OBJECT,
+            object_log.ObjectLogEntryType.RESTORE_OBJECT_VERSION,
+        }:
+            if entry.user_id not in creator_ids:
+                creator_ids.append(entry.user_id)
+            if entry.type in {
+                object_log.ObjectLogEntryType.CREATE_OBJECT,
+                object_log.ObjectLogEntryType.CREATE_BATCH
+            }:
+                creation_datetime = entry.utc_datetime
+                if version_id == 0:
[…]
