+++ docs/user_guide/export.rst
@@ -7,4 +7,4 @@ Export
 
 The PDF file consists of the :ref:`PDF files <pdf_export>` for the individual objects. As a result, only the current metadata versions and only those uploaded files that can be represented in the PDF are included.
 
-The archive contains a file ``data.json`` with information on objects in JSON format, including all metadata versions, comments, publications, location history and basic information on uploaded files, alongside information on relevant actions, instruments, users and locations. The uploaded files themselves are in subdirectories, grouped by the individual objects, so the first file uploaded for an object with the ID #42 would be located in the directory ``objects/42/files/0`` under its original upload name.
+The archive contains a file ``data.json`` with information on objects in JSON format, including all metadata versions, comments, publications, location history and basic information on uploaded files, alongside information on relevant actions, instruments, users and locations. The uploaded files themselves are in subdirectories, grouped by the individual objects, so the first file uploaded for an object with the ID #42 would be located in the directory ``objects/42/files/0`` under its original upload name. It also contains a ``.rdf`` file containing Dublin Core metadata in RDF/XML format for each object.
