### 2020-07-31 1799533
Allow downloading all files for an object as zip archive
