### 2020-08-19 187a506
Added fullscreen image preview for object and instrument log images
