### 2020-08-19 4c4117f
Added instrument log to HTTP API
