+++ docs/developer_guide/api.rst
@@ -704,6 +704,383 @@ Reading an instrument
     :statuscode 404: the instrument does not exist
 
 
+Instrument Log Entries
+----------------------
+
+Reading a list of all log entries for an instrument
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. http:get:: /api/v1/instruments/(int:instrument_id)/log_entries/
+
+    Get a list of all log entries for a specific instrument (`instrument_id`).
+
+    **Example request**:
+
+    .. sourcecode:: http
+
+        GET /api/v1/instruments/1/log_entries HTTP/1.1
+        Host: iffsamples.fz-juelich.de
+        Accept: application/json
+        Authorization: Basic dXNlcm5hbWU6cGFzc3dvcmQ=
+
+    **Example response**:
+
+    .. sourcecode:: http
+
+        HTTP/1.1 200 OK
+        Content-Type: application/json
+
+        [
+            {
+                "log_entry_id": 1,
+                "utc_datetime": "2020-08-19T12:13:14.123456",
+                "author": 1,
+                "content": "Example Log Entry 1",
+                "categories": []
+            },
+            {
+                "log_entry_id": 2,
+                "utc_datetime": "2020-08-19T13:14:15.123456",
+                "author": 1,
+                "content": "Example Log Entry 2",
+                "categories": [
+                    {
+                        "category_id": 1
+                        "title": "Error Report"
+                    },
+                    {
+                        "category_id": 7
+                        "title": "Maintenance Log"
+                    }
+                ]
+            }
+        ]
+
+    :statuscode 200: no error
+    :statuscode 403: the instrument log can only be accessed by instrument scientists
+    :statuscode 404: the instrument does not exist
+
+
+Reading an instrument log entry
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. http:get:: /api/v1/instruments/(int:instrument_id)/log_entries/(int:log_entry_id)
+
+    Get the specific log entry (`log_entry_id`) for an instrument (`instrument_id`).
+
+    **Example request**:
+
+    .. sourcecode:: http
+
+        GET /api/v1/instruments/1/log_entries/2 HTTP/1.1
+        Host: iffsamples.fz-juelich.de
+        Accept: application/json
+        Authorization: Basic dXNlcm5hbWU6cGFzc3dvcmQ=
+
+    **Example response**:
+
+    .. sourcecode:: http
+
+        HTTP/1.1 200 OK
+        Content-Type: application/json
+
+        {
+            "log_entry_id": 2,
+            "utc_datetime": "2020-08-19T13:14:15.123456",
+            "author": 1,
+            "content": "Example Log Entry 2",
+            "categories": [
+                {
+                    "category_id": 1
+                    "title": "Error Report"
+                },
+                {
+                    "category_id": 7
+                    "title": "Maintenance Log"
+                }
+            ]
+        }
+
+    :>json number log_entry_id: the log entry's ID
+    :>json string utc_datetime: the date and time of the log entry in UTC in ISO format
+    :>json string content: the log entry's content
+    :>json number author: the user ID of the log entry's author
+    :>json list categories: the log entry's categories
+    :statuscode 200: no error
+    :statuscode 403: the instrument log can only be accessed by instrument scientists
+    :statuscode 404: the instrument or the log entry do not exist
+
+
+Reading a list of all log categories for an instrument
[…]
