+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Currently in development.
 
 - Added Dublin Core metadata in RDF/XML format
 - Added fullscreen image preview of object and instrument log images
+- Added instrument log to HTTP API
 
 Version 0.12
 ------------
