### 2020-08-20 0e38405
Improve instrument log export
