### 2020-08-24 679d8bb
Add user affiliation
