+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Currently in development.
 - Added Dublin Core metadata in RDF/XML format
 - Added fullscreen image preview of object and instrument log images
 - Added instrument log to HTTP API
+- Allow setting a publicly visible user affiliation
 
 Version 0.12
 ------------
