### 2020-08-26 9c9d1b8
Add default setting for creating log entries during object creation
