+++ sampledb/logic/dataverse_export.py
@@ -1,0 +1,788 @@
+"""
+This module offers functions for exporting SampleDB information to a Dataverse
+instance with support for the Processing Metadata from the EngMeta schemas
+(see: https://www.izus.uni-stuttgart.de/fokus/engmeta).
+
+Some basic information is exported to the default Citation Metadata block and
+the process-specific metadata is exported to the Method Parameters fields of
+the Processing Metadata block.
+"""
+
+import datetime
+import json
+import typing
+from urllib.parse import urlencode
+
+import flask
+import requests
+
+from .. import db
+from .units import prettify_units
+from . import datatypes, object_log, users, actions, objects, errors, object_permissions, files, settings
+from ..models import DataverseExport
+
+
+DATAVERSE_TIMEOUT = 30
+
+
+def flatten_metadata(
+        metadata: typing.Union[typing.Dict[str, typing.Any], typing.List[typing.Any]],
+        path: typing.Optional[typing.List[typing.Union[str, int]]] = None
+) -> typing.Generator[typing.Tuple[typing.Dict[str, typing.Any], typing.List[typing.Union[str, int]]], None, None]:
+    """
+    Convert nested object data to a generator yielding the property data and path.
+
+    :param metadata: the nested object data
+    :param path: the path to start at, or None
+    :return: the generator for property data and path
+    """
+    if path is None:
+        path = []
+    if isinstance(metadata, list):
+        yield from _flatten_metadata_array(metadata, path)
+    elif isinstance(metadata, dict) and '_type' not in metadata:
+        yield from _flatten_metadata_object(metadata, path)
+    else:
+        yield metadata, path
+
+
+def _flatten_metadata_array(
+        metadata: typing.List[typing.Any],
+        path: typing.List[typing.Union[str, int]]
+) -> typing.Generator[typing.Tuple[typing.Dict[str, typing.Any], typing.List[typing.Union[str, int]]], None, None]:
+    for index, value in enumerate(metadata):
+        yield from flatten_metadata(value, path + [index])
+
+
+def _flatten_metadata_object(
+        metadata: typing.Dict[str, typing.Any],
+        path: typing.List[typing.Union[str, int]]
+) -> typing.Generator[typing.Tuple[typing.Dict[str, typing.Any], typing.List[typing.Union[str, int]]], None, None]:
+    for key, value in metadata.items():
+        yield from flatten_metadata(value, path + [key])
+
+
+def get_title_for_property(
+        path: typing.List[str],
+        schema: typing.Dict[str, typing.Any]
+) -> str:
+    """
+    Return the title path for a given property.
+
+    The title path consists of the individual titles or indices, joined by the
+    rightwards arrow symbol.
+
+    :param path: the path to a property
+    :param schema: the object schema
+    :return: the title path
+    """
+    try:
+        subschema = schema
+        title_path = []
+        for key in path:
+            if isinstance(key, str):
+                subschema = subschema['properties'][key]
+                title_path.append(subschema['title'])
+            else:
+                subschema = subschema['items']
+                title_path.append(key)
+    except Exception:
+        title_path = path
+    return ' → '.join(map(str, title_path))
+
+
+def get_property_export_default(
+        path: typing.List[str],
+        schema: typing.Dict[str, typing.Any]
+) -> bool:
+    """
+    Return whether dataverse_export is set to True in the schema of a property.
+
+    :param path: the path to a property
+    :param schema: the object schema
+    :return: the value of dataverse_export, or False if it is not set
+    """
+    if path == ['name']:
+        return True
+    try:
+        subschema = schema
+        for key in path:
+            if isinstance(key, str):
+                subschema = subschema['properties'][key]
+            else:
+                subschema = subschema['items']
[…]
