+++ sampledb/frontend/errors.py
@@ -6,7 +6,7 @@
 import flask
 
 from . import frontend
-from ..logic.errors import UserIsReadonlyError
+from ..logic.errors import UserIsReadonlyError, DataverseNotReachableError
 
 __author__ = 'Florian Rhiem <f.rhiem@fz-juelich.de>'
 
@@ -46,3 +46,8 @@ def internal_server_error(error):
 @frontend.errorhandler(UserIsReadonlyError)
 def user_is_readonly_error(error):
     return flask.render_template('errors/user_is_readonly.html'), 403
+
+
+@frontend.errorhandler(DataverseNotReachableError)
+def dataverse_not_reachable_error(error):
+    return flask.render_template('errors/dataverse_not_reachable.html'), 500
