+++ sampledb/logic/object_log.py
@@ -185,3 +185,14 @@ def reference_object_in_metadata(user_id: int, object_id: int, referencing_objec
             'object_id': referencing_object_id
         }
     )
+
+
+def export_to_dataverse(user_id: int, object_id: int, dataverse_url: str):
+    _store_new_log_entry(
+        type=ObjectLogEntryType.EXPORT_TO_DATAVERSE,
+        object_id=object_id,
+        user_id=user_id,
+        data={
+            'dataverse_url': dataverse_url
+        }
+    )
