+++ requirements.txt
@@ -60,6 +60,7 @@ pytz==2020.4
 qrcode==6.1
 reportlab==3.5.56
 requests==2.25.0
+requests-mock==1.8.0
 selenium==3.141.0
 six==1.15.0
 snowballstemmer==2.0.0
