+++ sampledb/version.py
@@ -5,4 +5,4 @@ This file should only contain the definition of the __version__ variable.
 It may be read by multiple other files, e.g. setup.py and the status page.
 """
 
-__version__ = '0.12.0'
+__version__ = '0.13.0'
