### 2020-09-03 98e4bf0
Use DOI for latest version instead of specific version
