+++ docs/user_guide/citations.rst
@@ -5,7 +5,7 @@ Citation Guide
 
 To cite SampleDB, please use:
 
-- Florian Rhiem et al., SampleDB. [Online; accessed <today>]. 2017-2020. DOI: 10.5281/zenodo.4012176. URL: https://github.com/sciapp/sampledb
+- Florian Rhiem et al., SampleDB. [Online; accessed <today>]. 2017-2020. DOI: 10.5281/zenodo.4012175. URL: https://github.com/sciapp/sampledb
 
 Here is an example BibTeX entry:
 
@@ -17,7 +17,7 @@ Here is an example BibTeX entry:
       year = {2017--2020},
       url = "https://github.com/sciapp/sampledb",
       note = {[Online; accessed 2020-09-02]},
-      doi = {10.5281/zenodo.4012176}
+      doi = {10.5281/zenodo.4012175}
     }
 
 If you wish to reference a specific sample, measurement or simulation in an instance of SampleDB, you can use its URL, e.g. ``https://iffsamples.fz-juelich.de/objects/0``. Bear in mind, however, that users need to be signed in even to view public objects.
