### 2020-09-03 d261f2a
Add Zenodo DOI badge to README.md
