### 2020-09-03 2f2d7c7
Add MIT license badge to README.md
