### 2020-09-09 5708652
Allow restricting location management to administrators
