### 2020-09-10 9c54e3d
Allow hiding actions
