### 2020-09-10 6b06af1
Add admin permissions setting
