### 2020-09-15 d657e68
Add object name to properties of publications linked to an object
