+++ sampledb/models/notifications.py
@@ -21,6 +21,7 @@ class NotificationType(enum.Enum):
     ANNOUNCEMENT = 4
     RECEIVED_OBJECT_PERMISSIONS_REQUEST = 5
     INSTRUMENT_LOG_ENTRY_CREATED = 6
+    REFERENCED_BY_OBJECT_METADATA = 7
 
 
 @enum.unique
