### 2020-09-15 da951fc
Make invitation time limit configurable
