### 2020-09-16 10da4e8
Show navbar dropdowns on hover
