### 2020-09-16 de3aec0
Allow copying permissions from another object during object creation
