+++ sampledb/frontend/objects.py
@@ -34,7 +34,7 @@ from ..logic.projects import get_project, get_user_projects, get_user_project_pe
 from ..logic.locations import get_location, get_object_ids_at_location, get_object_location_assignment, get_object_location_assignments, get_locations, assign_location_to_object, get_locations_tree
 from ..logic.files import FileLogEntryType
 from ..logic.errors import GroupDoesNotExistError, ObjectDoesNotExistError, UserDoesNotExistError, ActionDoesNotExistError, ValidationError, ProjectDoesNotExistError, LocationDoesNotExistError
-from .objects_forms import ObjectPermissionsForm, ObjectForm, ObjectVersionRestoreForm, ObjectUserPermissionsForm, CommentForm, ObjectGroupPermissionsForm, ObjectProjectPermissionsForm, FileForm, FileInformationForm, FileHidingForm, ObjectLocationAssignmentForm, ExternalLinkForm, ObjectPublicationForm
+from .objects_forms import ObjectPermissionsForm, ObjectForm, ObjectVersionRestoreForm, ObjectUserPermissionsForm, CommentForm, ObjectGroupPermissionsForm, ObjectProjectPermissionsForm, FileForm, FileInformationForm, FileHidingForm, ObjectLocationAssignmentForm, ExternalLinkForm, ObjectPublicationForm, CopyPermissionsForm
 from ..utils import object_permissions_required
 from .utils import jinja_filter, generate_qrcode
 from .object_form_parser import parse_form_data
@@ -535,12 +535,25 @@ def show_object_form(object, action, previous_object=None, should_upgrade_schema
         create_log_entry_default = None
         may_create_log_entry = False
 
+    if object is None and 'copy_permissions_from_other_object' in flask.request.form and 'copy_permissions_object_id' in flask.request.form:
+        copy_permissions_object_id = flask.request.form.get('copy_permissions_object_id')
+        try:
+            copy_permissions_object_id = int(copy_permissions_object_id)
+            if Permissions.READ not in get_user_object_permissions(copy_permissions_object_id, flask_login.current_user.id):
+                copy_permissions_object_id = None
+        except Exception:
+            pass
+    else:
+        copy_permissions_object_id = None
+
     if previous_object is not None:
         action_id = previous_object.action_id
         previous_object_id = previous_object.id
+        has_grant_for_previous_object = Permissions.GRANT in get_user_object_permissions(user_id=flask_login.current_user.id, object_id=previous_object_id)
     else:
         action_id = action.id
         previous_object_id = None
+        has_grant_for_previous_object = False
     errors = []
     object_errors = {}
     form_data = {}
@@ -619,7 +632,7 @@ def show_object_form(object, action, previous_object=None, should_upgrade_schema
                                 data_sequence.append(deepcopy(object_data))
                         else:
                             data_sequence = [object_data] * num_objects_in_batch
-                        objects = create_object_batch(action_id=action.id, data_sequence=data_sequence, user_id=flask_login.current_user.id)
+                        objects = create_object_batch(action_id=action.id, data_sequence=data_sequence, user_id=flask_login.current_user.id, copy_permissions_object_id=copy_permissions_object_id)
                         object_ids = [object.id for object in objects]
                         if category_ids is not None:
                             log_entry = logic.instrument_log_entries.create_instrument_log_entry(
@@ -636,7 +649,7 @@ def show_object_form(object, action, previous_object=None, should_upgrade_schema
                         flask.flash('The objects were created successfully.', 'success')
                         return flask.redirect(flask.url_for('.objects', ids=','.join([str(object_id) for object_id in object_ids])))
                     else:
-                        object = create_object(action_id=action.id, data=object_data, user_id=flask_login.current_user.id, previous_object_id=previous_object_id, schema=previous_object_schema)
[…]
