### 2020-09-16 b42065f
Improved invitation token handling
