+++ tests/logic/test_send_confirmation_email.py
@@ -16,7 +16,7 @@ def test_send_confirm_email(app):
 
         # email authentication for ldap-user
         with sampledb.mail.record_messages() as outbox:
-            sampledb.logic.utils.send_confirm_email(user.email, user.id, 'add_login')
+            sampledb.logic.utils.send_email_confirmation_email(user.email, user.id, 'add_login')
 
         assert len(outbox) == 1
         assert sampledb.logic.ldap.create_user_from_ldap(username).email in outbox[0].recipients
@@ -25,20 +25,19 @@ def test_send_confirm_email(app):
 
         # email invitation new user
         with sampledb.mail.record_messages() as outbox:
-            sampledb.logic.utils.send_confirm_email('test@fz-juelich.de', None, 'invitation')
+            sampledb.logic.utils.send_user_invitation_email('test@example.com', 0)
 
         assert len(outbox) == 1
-        assert 'test@fz-juelich.de' in outbox[0].recipients
+        assert 'test@example.com' in outbox[0].recipients
         message = outbox[0].html
-        assert 'SampleDB Email Confirmation' in message
+        assert 'SampleDB Invitation' in message
 
         # confirm new contact_email
         with sampledb.mail.record_messages() as outbox:
-            sampledb.logic.utils.send_confirm_email('testmail@fz-juelich.de', 1, 'edit_profile')
+            sampledb.logic.utils.send_email_confirmation_email('testmail@example.com', 1, 'edit_profile')
 
         assert len(outbox) == 1
-        assert 'testmail@fz-juelich.de' in outbox[0].recipients
+        assert 'testmail@example.com' in outbox[0].recipients
         message = outbox[0].html
         assert 'SampleDB Email Confirmation' in message
     app.config['SERVER_NAME'] = server_name
-
