### 2020-09-17 7421d7c
Avoid rebuilding flask server and app for each test
