+++ tests/scripts/test_export_action_schema.py
@@ -10,7 +10,6 @@ import tempfile
 from sampledb import db
 from sampledb.logic import instruments, actions
 import sampledb.__main__ as scripts
-from ..test_utils import app_context
 
 
 @pytest.fixture
