+++ tests/frontend/test_preferences.py
@@ -5,7 +5,6 @@
 
 import requests
 import pytest
-import bcrypt
 from bs4 import BeautifulSoup
 
 import sampledb
@@ -13,9 +12,6 @@ import sampledb.models
 import sampledb.logic
 from sampledb.logic.authentication import add_email_authentication
 from sampledb.logic import object_permissions, groups, projects
-
-
-from tests.test_utils import flask_server, app
 
 
 @pytest.fixture
