+++ docs/utils/generate_images.py
@@ -23,7 +23,6 @@ import sampledb
 import sampledb.config
 import sampledb.logic
 import tests.conftest
-import tests.test_utils
 
 
 def scroll_to(driver, x, y):
@@ -375,7 +374,7 @@ temp_dir = tempfile.mkdtemp()
 try:
     os.mkdir(os.path.join(temp_dir, 'uploaded_files'))
     sampledb.config.FILE_STORAGE_PATH = os.path.join(temp_dir, 'uploaded_files')
-    app = tests.test_utils.create_app()
+    app = tests.conftest.create_app()
     with app.app_context():
         user = sampledb.models.User(
             name="Example User",
@@ -465,7 +464,7 @@ try:
         # disable Chrome sandbox for root in GitLab CI
         if 'CI' in os.environ and getpass.getuser() == 'root':
             options.add_argument('--no-sandbox')
-        with contextlib.contextmanager(tests.test_utils.create_flask_server)(app) as flask_server:
+        with contextlib.contextmanager(tests.conftest.create_flask_server)(app) as flask_server:
             with contextlib.closing(Chrome(options=options)) as driver:
                 time.sleep(5)
                 object_permissions(flask_server.base_url, driver)
