### 2020-09-18 fb9c3da
Allow hiding instruments
