+++ docs/developer_guide/api.rst
@@ -659,6 +659,7 @@ Reading a list of all instruments
                 "instrument_id": 1,
                 "name": "Example Instrument",
                 "description": "This is an example instrument",
+                "is_hidden": false,
                 "instrument_scientists": [1, 42]
             }
         ]
@@ -693,12 +694,14 @@ Reading an instrument
             "instrument_id": 1,
             "name": "Example Instrument",
             "description": "This is an example instrument",
+            "is_hidden": false,
             "instrument_scientists": [1, 42]
         }
 
     :>json number instrument_id: the instrument's ID
     :>json string name: the instruments's name
     :>json string description: the instruments's description
+    :>json bool is_hidden: whether or not the instrument is hidden
     :>json list instrument_scientists: the instrument scientists' IDs
     :statuscode 200: no error
     :statuscode 404: the instrument does not exist
