### 2020-09-24 138cfc9
Only show Use as Template button if user may use the action for the object
