### 2020-09-25 f6504ff
Add versioning to instrument log entries
