+++ sampledb/models/instrument_log_entries.py
@@ -15,10 +15,15 @@ from .objects import Objects
 
 
 instrument_log_entry_category_association_table = db.Table(
-    'instrument_log_entry_category_associations',
+    'instrument_log_entry_version_category_associations',
     db.metadata,
-    db.Column('log_entry_id', db.Integer, db.ForeignKey('instrument_log_entries.id')),
-    db.Column('category_id', db.Integer, db.ForeignKey('instrument_log_categories.id', ondelete="CASCADE"))
+    db.Column('log_entry_id', db.Integer),
+    db.Column('log_entry_version_id', db.Integer),
+    db.Column('category_id', db.Integer, db.ForeignKey('instrument_log_categories.id', ondelete="CASCADE")),
+    db.ForeignKeyConstraint(
+        ['log_entry_id', 'log_entry_version_id'],
+        ['instrument_log_entry_versions.log_entry_id', 'instrument_log_entry_versions.version_id']
+    )
 )
 
 
@@ -45,7 +50,7 @@ class InstrumentLogCategory(db.Model):
         self.theme = theme
 
     def __repr__(self):
-        return '<{0}(id={1.id}, instrument_id={1.instrument_id}, title="{1.title}", theme={1.theme.name.lower()})>'.format(type(self).__name__, self)
+        return f'<{type(self).__name__}(id={self.id}, instrument_id={self.instrument_id}, title="{self.title}", theme={self.theme.name.lower()})>'
 
 
 class InstrumentLogEntry(db.Model):
@@ -54,27 +59,46 @@ class InstrumentLogEntry(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     instrument_id = db.Column(db.Integer, db.ForeignKey(Instrument.id), nullable=False)
     user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
-    content = db.Column(db.Text, nullable=False)
-    utc_datetime = db.Column(db.DateTime, nullable=False)
     author = db.relationship('User')
-    categories = db.relationship('InstrumentLogCategory', secondary=instrument_log_entry_category_association_table)
 
     def __init__(
             self,
             instrument_id: int,
-            user_id: int,
-            content: str,
-            utc_datetime: typing.Optional[datetime.datetime] = None
+            user_id: int
     ):
         self.instrument_id = instrument_id
         self.user_id = user_id
+
+    def __repr__(self):
+        return '<{0}(id={1.id}, instrument_id={1.instrument_id}, user_id={1.user_id})>'.format(type(self).__name__, self)
+
+
+class InstrumentLogEntryVersion(db.Model):
+    __tablename__ = 'instrument_log_entry_versions'
+
+    log_entry_id = db.Column(db.Integer, db.ForeignKey(InstrumentLogEntry.id), primary_key=True)
+    version_id = db.Column(db.Integer, primary_key=True)
+    content = db.Column(db.Text, nullable=False)
+    utc_datetime = db.Column(db.DateTime, nullable=False)
+    categories = db.relationship('InstrumentLogCategory', secondary=instrument_log_entry_category_association_table)
+    log_entry = db.relationship(InstrumentLogEntry, backref='versions')
+
+    def __init__(
+            self,
+            log_entry_id: int,
+            version_id: int,
+            content: str,
+            utc_datetime: typing.Optional[datetime.datetime] = None
+    ):
+        self.log_entry_id = log_entry_id
+        self.version_id = version_id
         self.content = content
         if utc_datetime is None:
             utc_datetime = datetime.datetime.utcnow()
         self.utc_datetime = utc_datetime
 
[…]
