+++ sampledb/logic/groups.py
@@ -14,12 +14,13 @@ SQLAlchemy ORM class.
 As the group models use flask-sqlalchemy however, the functions in this
 module should be called from within a Flask application context.
 """
-
-from sqlalchemy.exc import IntegrityError
 import collections
 import datetime
 import typing
+
 import flask
+from sqlalchemy.exc import IntegrityError
+
 from .. import db
 from ..models import groups
 from .users import get_user
