+++ sampledb/models/notifications.py
@@ -22,6 +22,7 @@ class NotificationType(enum.Enum):
     RECEIVED_OBJECT_PERMISSIONS_REQUEST = 5
     INSTRUMENT_LOG_ENTRY_CREATED = 6
     REFERENCED_BY_OBJECT_METADATA = 7
+    INSTRUMENT_LOG_ENTRY_EDITED = 8
 
 
 @enum.unique
