### 2020-09-30 56b8a3a
Allow administrators to deactivate users
