### 2020-10-01 3d52466
Allow custom action types
