+++ tests/frontend/test_favorites.py
@@ -27,7 +27,7 @@ def user_session(flask_server):
 @pytest.fixture
 def actions(flask_server):
     actions = [sampledb.models.Action(
-        action_type=sampledb.models.ActionType.SAMPLE_CREATION,
+        action_type_id=sampledb.models.ActionType.SAMPLE_CREATION,
         name='Action {}'.format(i + 1),
         schema={
             'title': 'Example Object',
