+++ sampledb/logic/schemas/validate_schema.py
@@ -54,6 +54,8 @@ def validate_schema(schema: dict, path: typing.Optional[typing.List[str]] = None
         return _validate_sample_schema(schema, path)
     elif schema['type'] == 'measurement':
         return _validate_measurement_schema(schema, path)
+    elif schema['type'] == 'object_reference':
+        return _validate_object_reference_schema(schema, path)
     elif schema['type'] == 'tags':
         return _validate_tags_schema(schema, path)
     elif schema['type'] == 'hazards':
@@ -409,6 +411,29 @@ def _validate_measurement_schema(schema: dict, path: typing.List[str]) -> None:
         raise ValidationError('missing keys in schema: {}'.format(missing_keys), path)
     if 'note' in schema and not isinstance(schema['note'], str):
         raise ValidationError('note must be str', path)
+
+
+def _validate_object_reference_schema(schema: dict, path: typing.List[str]) -> None:
+    """
+    Validates the given object reference object schema and raises a ValidationError if it is invalid.
+
+    :param schema: the sampledb object schema
+    :param path: the path to this subschema
+    :raise ValidationError: if the schema is invalid.
+    """
+    valid_keys = {'type', 'title', 'note', 'action_type_id'}
+    required_keys = {'type', 'title'}
+    schema_keys = set(schema.keys())
+    invalid_keys = schema_keys - valid_keys
+    if invalid_keys:
+        raise ValidationError('unexpected keys in schema: {}'.format(invalid_keys), path)
+    missing_keys = required_keys - schema_keys
+    if missing_keys:
+        raise ValidationError('missing keys in schema: {}'.format(missing_keys), path)
+    if 'note' in schema and not isinstance(schema['note'], str):
+        raise ValidationError('note must be str', path)
+    if 'action_type_id' in schema and not isinstance(schema['action_type_id'], (int, type(None))):
+        raise ValidationError('action_type_id must be int or None', path)
 
 
 def _validate_notebook_templates(notebook_templates: typing.Any) -> None:
