### 2020-10-01 8d8b874
Allow setting instrument log entry order
