+++ sampledb/logic/settings.py
@@ -26,6 +26,7 @@ DEFAULT_SETTINGS = {
     "USE_SCHEMA_EDITOR": True,
     "USE_ADMIN_PERMISSIONS": False,
     "SHOW_INVITATION_LOG": False,
+    "INSTRUMENT_LOG_ORDER_ASCENDING": True
 }
 
 
