### 2020-10-01 8d979cc
Adding background loading of objects
