### 2020-10-01 ba67a87
Remove duplicate search for referencable objects
