### 2020-10-01 6a19c1c
Fix simulation default type plural object name
