### 2020-10-01 ebdba1b
Adding loading to form_table_*, adding test for new endpoint
