### 2020-10-02 be54147
Adding perm level argument to objects/references and use it for permission copy forms Updated changelog and docs Making js compatible with more browsers Adding visual representation for loading
