+++ sampledb/logic/object_permissions.py
@@ -274,6 +274,7 @@ def get_objects_with_permissions(
         action_type_id: typing.Optional[int] = None,
         project_id: typing.Optional[int] = None,
         object_ids: typing.Optional[typing.Sequence[int]] = None,
+        num_objects_found: typing.Optional[typing.List[int]] = None,
         **kwargs
 ) -> typing.List[Object]:
 
@@ -325,20 +326,29 @@ def get_objects_with_permissions(
         'user_id': user_id
     }
 
-    objs = objects.get_objects(filter_func=filter_func, action_filter=action_filter, table=table, parameters=parameters, sorting_func=sorting_func, limit=limit, offset=offset, **kwargs)
-    if project_id is not None:
-        filtered_objs = []
-        for obj in objs:
-            project_object_permissions = ProjectObjectPermissions.query.filter_by(object_id=obj.object_id, project_id=project_id).first()
-            if project_object_permissions is not None and permissions in project_object_permissions.permissions:
-                filtered_objs.append(obj)
-        objs = filtered_objs
-    if object_ids is not None:
-        filtered_objs = []
-        for obj in objs:
-            if obj.object_id in object_ids:
-                filtered_objs.append(obj)
-        objs = filtered_objs
+    if project_id is None and object_ids is None:
+        objs = objects.get_objects(filter_func=filter_func, action_filter=action_filter, table=table, parameters=parameters, sorting_func=sorting_func, limit=limit, offset=offset, num_objects_found=num_objects_found, **kwargs)
+    else:
+        objs = objects.get_objects(filter_func=filter_func, action_filter=action_filter, table=table, parameters=parameters, sorting_func=sorting_func, **kwargs)
+        if project_id is not None:
+            filtered_objs = []
+            for obj in objs:
+                project_object_permissions = ProjectObjectPermissions.query.filter_by(object_id=obj.object_id, project_id=project_id).first()
+                if project_object_permissions is not None and permissions in project_object_permissions.permissions:
+                    filtered_objs.append(obj)
+            objs = filtered_objs
+        if object_ids is not None:
+            filtered_objs = []
+            for obj in objs:
+                if obj.object_id in object_ids:
+                    filtered_objs.append(obj)
+            objs = filtered_objs
+        if num_objects_found is not None:
+            num_objects_found.append(len(objs))
+        if offset:
+            objs = objs[offset:]
+        if limit:
+            objs = objs[:limit]
     return objs
 
 
