### 2020-10-06 2f5694d
Allow disabling group deletion by non-administrators
