+++ docs/administrator_guide/configuration.rst
@@ -150,5 +150,7 @@ Miscellaneous
      - The time limit for invitation links in seconds.
    * - SAMPLEDB_ONLY_ADMINS_CAN_MANAGE_LOCATIONS
      - If set, only administrators will be able to create and edit locations.
+   * - SAMPLEDB_ONLY_ADMINS_CAN_DELETE_GROUPS
+     - If set, only administrators will be able to delete non-empty groups.
 
 There are other configuration values related to packages used by SampleDB. For more information on those, see the documentation of the corresponding packages.
