### 2020-10-07 fb2dd2b
Add docker-compose.yml file and instructions
Also add a .dockerignore
