### 2020-10-07 5d1ce3a
Ensure that file storage path is owned by sampledb user in docker container
