### 2020-10-07 f11388c
fix incorrect mail_server env var
