### 2020-10-07 a1f9728
Add link to docker-compose.yml to README.md
