+++ README.md
@@ -19,7 +19,7 @@ If you do not have Docker installed yet, please [install Docker](https://docs.do
 
 ### Using docker-compose
 
-First, get the docker-compose.yml configuration file. You can git clone this repo or just get the file:
+First, get the [docker-compose.yml](https://raw.githubusercontent.com/sciapp/sampledb/develop/docker-compose.yml) configuration file. You can git clone this repo or just get the file:
 
 ```bash
 curl https://raw.githubusercontent.com/sciapp/sampledb/develop/docker-compose.yml --output docker-compose.yml
