### 2020-10-13 1bb401e
Add "list" array style
