+++ requirements.txt
@@ -29,7 +29,7 @@ Flask-WTF==0.14.3
 idna==2.10
 imagesize==1.2.0
 importlib-metadata==2.0.0
-iniconfig==1.0.1
+iniconfig==1.1.1
 itsdangerous==1.1.0
 jaraco.classes==3.1.0
 jaraco.collections==3.0.0
@@ -38,12 +38,12 @@ jaraco.text==3.2.0
 Jinja2==2.11.2
 jsonschema==3.2.0
 ldap3==2.8.1
-lxml==4.5.2
-Markdown==3.3
+lxml==4.6.1
+Markdown==3.3.2
 MarkupSafe==1.1.1
 more-itertools==8.5.0
 packaging==20.4
-Pillow==7.2.0
+Pillow==8.0.1
 Pint==0.16.1
 pluggy==0.13.1
 portend==2.6
@@ -51,7 +51,7 @@ psycopg2-binary==2.8.6
 py==1.9.0
 pyasn1==0.4.8
 pycparser==2.20
-Pygments==2.7.1
+Pygments==2.7.2
 pyparsing==2.4.7
 pyrsistent==0.17.3
 pytest==6.1.1
@@ -76,11 +76,11 @@ sphinxcontrib-websupport==1.2.4
 SQLAlchemy==1.3.20
 tempora==4.0.0
 toml==0.10.1
-urllib3==1.25.10
+urllib3==1.25.11
 vcversioner==2.16.0.0
 wcwidth==0.2.5
 webencodings==0.5.1
 Werkzeug==1.0.1
 WTForms==2.3.3
 zc.lockfile==2.0
-zipp==3.3.0
+zipp==3.4.0
