+++ docs/conf.py
@@ -100,7 +100,7 @@ epub_exclude_files = ['search.html']
 
 
 def setup(app):
-    app.add_stylesheet('css/custom.css')
+    app.add_css_file('css/custom.css')
 
 
 rst_prolog = """
