+++ requirements.txt
@@ -1,78 +1,86 @@
 alabaster==0.7.12
 aniso8601==8.0.0
-atomicwrites==1.3.0
-attrs==19.3.0
-Babel==2.7.0
+atomicwrites==1.4.0
+attrs==20.2.0
+Babel==2.8.0
 backports.functools-lru-cache==1.6.1
-bcrypt==3.1.7
-beautifulsoup4==4.8.1
-bleach==3.1.5
+bcrypt==3.2.0
+beautifulsoup4==4.9.3
+bleach==3.2.1
 blinker==1.4
-certifi==2019.9.11
-cffi==1.13.2
+certifi==2020.6.20
+cffi==1.14.3
 chardet==3.0.4
-cheroot==8.2.1
-CherryPy==18.4.0
+cheroot==8.4.5
+CherryPy==18.6.0
 chromedriver-binary==85.0.4183.87.0
-Click==7.0
-coverage==4.5.4
-docutils==0.15.2
-Flask==1.1.1
-Flask-HTTPAuth==3.3.0
-Flask-Login==0.4.1
+click==7.1.2
+coverage==5.3
+dnspython==2.0.0
+docutils==0.16
+email-validator==1.1.1
+Flask==1.1.2
+Flask-HTTPAuth==4.1.0
+Flask-Login==0.5.0
 Flask-Mail==0.9.1
-Flask-RESTful==0.3.7
-Flask-SQLAlchemy==2.4.1
-Flask-WTF==0.14.2
-idna==2.8
-imagesize==1.1.0
-importlib-metadata==0.23
+Flask-RESTful==0.3.8
+Flask-SQLAlchemy==2.4.4
+Flask-WTF==0.14.3
+idna==2.10
+imagesize==1.2.0
+importlib-metadata==2.0.0
+iniconfig==1.0.1
 itsdangerous==1.1.0
-jaraco.functools==2.0
-Jinja2==2.10.3
-jsonschema==3.1.1
-ldap3==2.6.1
+jaraco.classes==3.1.0
+jaraco.collections==3.0.0
+jaraco.functools==3.0.1
+jaraco.text==3.2.0
+Jinja2==2.11.2
+jsonschema==3.2.0
+ldap3==2.8.1
 lxml==4.5.2
-Markdown==3.2.2
+Markdown==3.3
 MarkupSafe==1.1.1
-more-itertools==7.2.0
-packaging==19.2
-Pillow==6.2.1
-Pint==0.9
-pluggy==0.13.0
+more-itertools==8.5.0
+packaging==20.4
+Pillow==7.2.0
+Pint==0.16.1
+pluggy==0.13.1
 portend==2.6
-psycopg2-binary==2.8.4
-py==1.8.0
-pyasn1==0.4.7
-pycparser==2.19
-Pygments==2.4.2
-pyparsing==2.4.5
-pyrsistent==0.15.5
-pytest==5.2.2
-pytest-cov==2.8.1
-pytz==2019.3
+psycopg2-binary==2.8.6
+py==1.9.0
+pyasn1==0.4.8
+pycparser==2.20
+Pygments==2.7.1
+pyparsing==2.4.7
+pyrsistent==0.17.3
+pytest==6.1.1
+pytest-cov==2.10.1
+pytz==2020.1
 qrcode==6.1
-reportlab==3.5.32
-requests==2.22.0
+reportlab==3.5.53
+requests==2.24.0
 selenium==3.141.0
-six==1.13.0
+six==1.15.0
 snowballstemmer==2.0.0
-soupsieve==1.9.5
-Sphinx==2.2.1
-sphinxcontrib-applehelp==1.0.1
[…]
