### 2020-11-02 6e37dee
Bugfix: Using dynamic web root in load objects js
