+++ README.md
@@ -62,7 +62,7 @@ docker run \
     --restart=always \
     --name sampledb \
     -p 8000:8000 \
-    sciapp/sampledb:0.14.1
+    sciapp/sampledb:0.15.0
 ```
 
 ### Once it's started
