### 2020-11-10 9a6deb6
Fix performance issues caused by instrument log object attachments
