### 2020-11-11 cc898b1
Improve performance of object lists
