+++ sampledb/logic/object_permissions.py
@@ -345,6 +345,7 @@ def get_objects_with_permissions(
         project_id: typing.Optional[int] = None,
         object_ids: typing.Optional[typing.Sequence[int]] = None,
         num_objects_found: typing.Optional[typing.List[int]] = None,
+        name_only: bool = False,
         **kwargs
 ) -> typing.List[Object]:
 
@@ -363,63 +364,75 @@ def get_objects_with_permissions(
     else:
         action_filter = None
 
-    if user.is_admin and settings.get_user_settings(user.id)['USE_ADMIN_PERMISSIONS']:
-        # admins who use admin permissions do not need permission-based filtering
-        table = None
-    else:
-        stmt = db.text("""
-        SELECT
-        o.object_id, o.version_id, o.action_id, o.data, o.schema, o.user_id, o.utc_datetime
-        FROM (
-            SELECT
-            object_id
-            FROM user_object_permissions_by_all
-            WHERE user_id = :user_id OR user_id IS NULL
-            GROUP BY (object_id)
-            HAVING MAX(permissions_int) >= :min_permissions_int
-        ) AS p
-        JOIN objects_current AS o ON o.object_id = p.object_id
-        """)
-        stmt = stmt.columns(
-            objects.Objects._current_table.c.object_id,
-            objects.Objects._current_table.c.version_id,
-            objects.Objects._current_table.c.action_id,
-            objects.Objects._current_table.c.data,
-            objects.Objects._current_table.c.schema,
-            objects.Objects._current_table.c.user_id,
-            objects.Objects._current_table.c.utc_datetime
-        )
-        table = sqlalchemy.sql.alias(stmt)
-
     parameters = {
         'min_permissions_int': permissions.value,
         'user_id': user_id
     }
 
-    if project_id is None and object_ids is None:
-        objs = objects.get_objects(filter_func=filter_func, action_filter=action_filter, table=table, parameters=parameters, sorting_func=sorting_func, limit=limit, offset=offset, num_objects_found=num_objects_found, **kwargs)
+    if name_only:
+        stmt = """
+        SELECT
+        o.object_id, o.version_id, o.action_id, jsonb_set('{"name": {"_type": "text", "text": ""}}', '{name,text}', o.data -> 'name' -> 'text') as data, '{"title": "Object", "type": "object", "properties": {"name": {"title": "Name", "type": "text"}}}'::jsonb as schema, o.user_id, o.utc_datetime
+        FROM objects_current AS o
+        """
     else:
-        objs = objects.get_objects(filter_func=filter_func, action_filter=action_filter, table=table, parameters=parameters, sorting_func=sorting_func, **kwargs)
-        if project_id is not None:
-            filtered_objs = []
-            for obj in objs:
-                project_object_permissions = ProjectObjectPermissions.query.filter_by(object_id=obj.object_id, project_id=project_id).first()
-                if project_object_permissions is not None and permissions in project_object_permissions.permissions:
-                    filtered_objs.append(obj)
-            objs = filtered_objs
-        if object_ids is not None:
-            filtered_objs = []
-            for obj in objs:
-                if obj.object_id in object_ids:
-                    filtered_objs.append(obj)
-            objs = filtered_objs
-        if num_objects_found is not None:
-            num_objects_found.append(len(objs))
-        if offset:
-            objs = objs[offset:]
-        if limit:
-            objs = objs[:limit]
-    return objs
+        stmt = """
+        SELECT
+        o.object_id, o.version_id, o.action_id, o.data, o.schema, o.user_id, o.utc_datetime
+        FROM objects_current AS o
+        """
+
[…]
