### 2020-11-12 9490015
Fix missing check for user None
