+++ sampledb/api/server/authentication.py
@@ -34,7 +34,7 @@ def verify_password(username, password):
     if not username:
         return None
     user = login(username, password)
-    if not user.is_active:
+    if user is None or not user.is_active:
         return None
     flask.g.user = user
     return user
