### 2020-11-13 9a8a72d
Fix missing check for user None for api tokens
