### 2020-11-18 a613e93
Add support for placeholder texts for text and quantity schemas
