+++ tests/logic/schemas/test_validate_schema.py
@@ -386,6 +386,36 @@ def test_validate_text_valid_multiline_type():
     validate_schema(wrap_into_basic_schema(schema))
 
 
+def test_validate_text_with_placeholder():
+    schema = {
+        'title': 'Example',
+        'type': 'text',
+        'placeholder': 'Placeholder'
+    }
+    validate_schema(wrap_into_basic_schema(schema))
+
+
+def test_validate_text_with_invalid_placeholder():
+    schema = {
+        'title': 'Example',
+        'type': 'text',
+        'placeholder': 1
+    }
+    with pytest.raises(ValidationError):
+        validate_schema(wrap_into_basic_schema(schema))
+
+
+def test_validate_text_with_placeholder_and_choices():
+    schema = {
+        'title': 'Example',
+        'type': 'text',
+        'placeholder': 'Placeholder',
+        'choices': ['A', 'B', 'C']
+    }
+    with pytest.raises(ValidationError):
+        validate_schema(wrap_into_basic_schema(schema))
+
+
 def test_validate_datetime_schema():
     schema = {
         'title': 'Example',
@@ -538,6 +568,27 @@ def test_validate_quantity_schema_missing_key():
     schema = {
         'title': 'Example',
         'type': 'quantity'
+    }
+    with pytest.raises(ValidationError):
+        validate_schema(wrap_into_basic_schema(schema))
+
+
+def test_validate_quantity_schema_with_placeholder():
+    schema = {
+        'title': 'Example',
+        'type': 'quantity',
+        'units': 'm',
+        'placeholder': 'Placeholder'
+    }
+    validate_schema(wrap_into_basic_schema(schema))
+
+
+def test_validate_quantity_schema_with_invalid_placeholder():
+    schema = {
+        'title': 'Example',
+        'type': 'quantity',
+        'units': 'm',
+        'placeholder': 1
     }
     with pytest.raises(ValidationError):
         validate_schema(wrap_into_basic_schema(schema))
