### 2020-11-19 d341a82
Add limit, offset and name_only option to HTTP API objects endpoint
