### 2020-11-25 d2bf1a0
Display projects based on parent-child relationship
