### 2020-11-25 3d9b544
Allow setting display properties as part of the object list URL
