### 2020-11-27 231896c
Improve "View Objects" for users, groups and projects
