### 2020-12-01 8e6cf21
Features: * Allowing partial object updates if the new config variable ALLOW_PARTIAL_OBJECT_UPDATES is true * Automatically assume that the unit of a Quantity matches the scheme unit if not given * Accept either magnitude or magnitude_in_base_units or both for a Quantity object. If both are given they have to match each other, otherwise automatically calculate the other one * Return both magnitude and magnitude_in_base_units Additional fixes: * Boolean config variables from the env were represented as a string despite the default value being set as an actual boolean * The quantity validation f
[…]
