+++ requirements.txt
@@ -1,5 +1,5 @@
 alabaster==0.7.12
-aniso8601==8.0.0
+aniso8601==8.1.0
 atomicwrites==1.4.0
 attrs==20.3.0
 Babel==2.9.0
@@ -9,18 +9,18 @@ beautifulsoup4==4.9.3
 bleach==3.2.1
 blinker==1.4
 certifi==2020.11.8
-cffi==1.14.3
+cffi==1.14.4
 chardet==3.0.4
-cheroot==8.4.7
+cheroot==8.4.8
 CherryPy==18.6.0
-chromedriver-binary==85.0.4183.87.0
+chromedriver-binary-auto==0.1
 click==7.1.2
 coverage==5.3
 dnspython==2.0.0
 docutils==0.16
 email-validator==1.1.2
 Flask==1.1.2
-Flask-HTTPAuth==4.1.0
+Flask-HTTPAuth==4.2.0
 Flask-Login==0.5.0
 Flask-Mail==0.9.1
 Flask-RESTful==0.3.8
@@ -28,7 +28,7 @@ Flask-SQLAlchemy==2.4.4
 Flask-WTF==0.14.3
 idna==2.10
 imagesize==1.2.0
-importlib-metadata==2.0.0
+importlib-metadata==3.1.1
 iniconfig==1.1.1
 itsdangerous==1.1.0
 jaraco.classes==3.1.0
@@ -38,15 +38,15 @@ jaraco.text==3.2.0
 Jinja2==2.11.2
 jsonschema==3.2.0
 ldap3==2.8.1
-lxml==4.6.1
+lxml==4.6.2
 Markdown==3.3.3
 MarkupSafe==1.1.1
 more-itertools==8.6.0
-packaging==20.4
+packaging==20.7
 Pillow==8.0.1
 Pint==0.16.1
 pluggy==0.13.1
-portend==2.6
+portend==2.7
 psycopg2-binary==2.8.6
 py==1.9.0
 pyasn1==0.4.8
@@ -58,7 +58,7 @@ pytest==6.1.2
 pytest-cov==2.10.1
 pytz==2020.4
 qrcode==6.1
-reportlab==3.5.55
+reportlab==3.5.56
 requests==2.25.0
 selenium==3.141.0
 six==1.15.0
