### 2020-12-08 edc2b2d
Added favicon
