### 2020-12-08 6187b2f
Run tests in parallel on GitLab CI
