+++ tests/scripts/test_list_actions.py
@@ -35,8 +35,8 @@ def actions(instrument):
 def test_list_actions(instrument, actions, capsys):
     scripts.main([scripts.__file__, 'list_actions'])
     output = capsys.readouterr()[0]
-    for action_id, action_name in [(1, 'Action 1'), (2, 'Action 2')]:
-        assert '- #{0}: {1}'.format(action_id, action_name) in output
+    for action in sampledb.logic.actions.get_actions():
+        assert '- #{0}: {1}'.format(action.id, action.name) in output
 
 
 def test_list_actions_arguments(instrument, actions, capsys):
