### 2020-12-10 0dbc5ec
Add cache to markdown_to_safe_html
