+++ sampledb/models/__init__.py
@@ -11,6 +11,7 @@ from . import groups
 from . import instruments
 from . import instrument_log_entries
 from . import locations
+from . import markdown_to_html_cache
 from . import markdown_images
 from . import objects
 from . import object_permissions
@@ -30,6 +31,7 @@ from .groups import Group
 from .instruments import Instrument
 from .instrument_log_entries import InstrumentLogEntry
 from .locations import Location, ObjectLocationAssignment
+from .markdown_to_html_cache import MarkdownToHTMLCacheEntry
 from .markdown_images import MarkdownImage
 from .notifications import Notification, NotificationType, NotificationMode, NotificationModeForType
 from .objects import Objects, Object
@@ -54,6 +56,7 @@ __all__ = [
     'instruments',
     'instrument_log_entries',
     'locations',
+    'markdown_to_html_cache',
     'markdown_images',
     'objects',
     'object_permissions',
@@ -80,6 +83,7 @@ __all__ = [
     'InstrumentLogEntry',
     'Location',
     'ObjectLocationAssignment',
+    'MarkdownToHTMLCacheEntry',
     'MarkdownImage',
     'Notification',
     'NotificationType',
