+++ sampledb/logic/utils.py
@@ -4,6 +4,7 @@
 """
 
 import smtplib
+
 import flask
 import flask_mail
 
