### 2020-12-10 ab8861d
Do not store markdown descriptions and notes directly as HTML
