### 2020-12-10 ff25061
Add short descriptions to actions and instruments
