+++ sampledb/models/actions.py
@@ -55,8 +55,22 @@ class Action(db.Model):
     user = db.relationship("User", backref="actions")
     description_is_markdown = db.Column(db.Boolean, nullable=False, default=False)
     is_hidden = db.Column(db.Boolean, nullable=False, default=False)
+    short_description = db.Column(db.String, nullable=False, default='')
+    short_description_is_markdown = db.Column(db.Boolean, nullable=False, default=False)
 
-    def __init__(self, action_type_id: int, name: str, schema: dict, description: str = '', instrument_id: typing.Optional[int] = None, user_id: typing.Optional[int] = None, description_is_markdown: bool = False, is_hidden: bool = False):
+    def __init__(
+            self,
+            action_type_id: int,
+            name: str,
+            schema: dict,
+            description: str = '',
+            instrument_id: typing.Optional[int] = None,
+            user_id: typing.Optional[int] = None,
+            description_is_markdown: bool = False,
+            is_hidden: bool = False,
+            short_description: str = '',
+            short_description_is_markdown: bool = False
+    ):
         self.type_id = action_type_id
         self.name = name
         self.description = description
@@ -65,6 +79,8 @@ class Action(db.Model):
         self.user_id = user_id
         self.description_is_markdown = description_is_markdown
         self.is_hidden = is_hidden
+        self.short_description = short_description
+        self.short_description_is_markdown = short_description_is_markdown
 
     def __eq__(self, other):
         return (
@@ -74,6 +90,8 @@ class Action(db.Model):
             self.name == other.name and
             self.description == other.description and
             self.description_is_markdown == other.description_is_markdown and
+            self.short_description == other.short_description and
+            self.short_description_is_markdown == other.short_description_is_markdown and
             self.is_hidden == other.is_hidden and
             self.schema == other.schema
         )
