+++ tests/logic/test_export.py
@@ -169,7 +169,9 @@ def validate_data(data):
                 'user_id': None,
                 'instrument_id': None,
                 'description': '',
-                'description_is_markdown': False
+                'description_is_markdown': False,
+                'short_description': '',
+                'short_description_is_markdown': False
             }
         ],
         'instruments': [
@@ -178,6 +180,8 @@ def validate_data(data):
                 'name': 'Example Instrument',
                 'description': 'Example Instrument Description',
                 'description_is_markdown': False,
+                'short_description': '',
+                'short_description_is_markdown': False,
                 'instrument_scientist_ids': [],
                 'instrument_log_entries': [
                     {
