### 2020-12-11 0d42f31
Fix search for array and non-array attributes with the same name
