+++ sampledb/static/js/schema_editor.js
@@ -247,7 +247,7 @@ $(function() {
       help_block.html("Objects must have a property 'name' as a <i>Text (Simple)</i> property.");
       help_parent.addClass("has-error");
       has_error = true;
-    } else if (!('type' in schema['properties']['name']) || schema['properties']['name']['type'] !== "text" || ('multiline' in schema['properties']['name'] && schema['properties']['name']['multiline']) || 'choices' in schema['properties']['name']) {
+    } else if (!('type' in schema['properties']['name']) || schema['properties']['name']['type'] !== "text" || ('multiline' in schema['properties']['name'] && schema['properties']['name']['multiline']) || ('markdown' in schema['properties']['name'] && schema['properties']['name']['markdown']) || 'choices' in schema['properties']['name']) {
       help_block.html("Object name must be a <i>Text (Simple)</i> property.");
       help_parent.addClass("has-error");
       has_error = true;
@@ -289,6 +289,8 @@ $(function() {
     if (schema['type'] === 'text') {
       if ('choices' in schema) {
         type = "choice";
+      } else if ('markdown' in schema && schema['markdown']) {
+        type = "markdown";
       } else if ('multiline' in schema && schema['multiline']) {
         type = "multiline";
       } else {
@@ -332,6 +334,8 @@ $(function() {
         updateChoiceProperty(path, real_path);
       } else if (type === "multiline") {
         updateMultilineProperty(path, real_path);
+      } else if (type === "markdown") {
+        updateMarkdownProperty(path, real_path);
       } else if (type === "bool") {
         updateBoolProperty(path, real_path);
       } else if (type === "sample") {
@@ -785,6 +789,104 @@ $(function() {
 
       var has_maxlength = $('#schema-editor-object__' + path.join('__') + '-multiline-maxlength-checkbox').prop('checked');
       var maxlength_input = $('#schema-editor-object__' + path.join('__') + '-multiline-maxlength-input');
+      var maxlength = maxlength_input.val();
+      var maxlength_group = maxlength_input.parent();
+      var maxlength_help = maxlength_group.find('.help-block');
+
+      if (has_maxlength) {
+        if (isNaN(maxlength)) {
+          maxlength_help.text("Please enter a number.");
+          maxlength_group.addClass("has-error");
+          has_error = true;
+        } else {
+          maxlength = Number.parseInt(maxlength);
+          if (maxlength >= 0) {
+            property_schema['maxLength'] = maxlength;
+            maxlength_help.text("");
+            maxlength_group.removeClass("has-error");
+          } else {
+            maxlength_help.text("Please enter a number greater than or equal to zero.");
+            maxlength_group.addClass("has-error");
+            has_error = true;
+          }
+        }
+      } else {
+        delete property_schema["maxLength"];
+        maxlength_help.text("");
+        maxlength_group.removeClass("has-error");
+      }
+      if ('minLength' in property_schema && 'maxLength' in property_schema && property_schema['minLength'] > property_schema['maxLength']) {
+        minlength_help.text("Please enter a number less than or equal to the maximum length.");
+        minlength_group.addClass("has-error");
+        maxlength_help.text("Please enter a number greater than or equal to the minimum length.");
+        maxlength_group.addClass("has-error");
+        has_error = true;
+      }
+
+      schema['properties'][real_path[real_path.length-1]] = property_schema;
+
+      input_schema.text(JSON.stringify(schema, null, 4));
+
+      window.schema_editor_errors[path.join('__') + '__specific'] = true;
+      if (!has_error) {
+        delete window.schema_editor_errors[path.join('__') + '__specific'];
+      }
[…]
