+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Currently in development.
 - Added short descriptions to actions and instruments
 - Allow linking to headers in Markdown content
 - Allow disabling of "Use in Measurement" button for samples
+- Added markdown support to object metadata
 
 Version 0.16
 ------------
