### 2020-12-14 38a3cdd
Allow linking to headers in Markdown content
