+++ sampledb/frontend/pdfexport.py
@@ -2,25 +2,25 @@
 import base64
 import datetime
 import io
-import json
+import os
 import typing
+import urllib.parse
 
 import flask
 import flask_login
-from reportlab.pdfgen.canvas import Canvas
-from reportlab.pdfbase.pdfdoc import PDFInfo
-from reportlab.lib.pagesizes import A4 as pagesize
-from reportlab.lib.units import mm
-from PIL import Image
+import jinja2
 import qrcode
 import qrcode.image.pil
+from weasyprint import default_url_fetcher, HTML
 
 from .. import logic
 from ..logic import object_log
 from ..logic.objects import get_object
-from ..logic.actions import get_action
 from ..logic.object_log import ObjectLogEntryType
 from ..logic.users import get_user
+
+from .markdown_images import IMAGE_FORMATS
+from .objects import get_object_if_current_user_has_read_permissions
 
 SECTIONS = {
     'activity_log',
@@ -30,539 +30,202 @@ SECTIONS = {
     'comments'
 }
 
-PAGE_WIDTH, PAGE_HEIGHT = pagesize
-LEFT_MARGIN = 25.4 * mm
-RIGHT_MARGIN = 25.4 * mm
-TOP_MARGIN = 31.7 * mm
-BOTTOM_MARGIN = 31.7 * mm
-
-
-def _set_up_page(canvas, object_id, qrcode_uri, logo_uri=None, logo_aspect_ratio=1, logo_alignment='right'):
-    canvas.saveState()
-    canvas.drawImage(qrcode_uri, PAGE_WIDTH - 27.5 * mm, PAGE_HEIGHT - 27.5 * mm, 20 * mm, 20 * mm)
-    if logo_uri:
-        if logo_aspect_ratio > 1:
-            logo_width = 19 * mm * logo_aspect_ratio
-            logo_height = 19 * mm
-        else:
-            logo_width = 19 * mm
-            logo_height = 19 * mm / logo_aspect_ratio
-        # right margin due to QR code
-        right_margin = 27.5 * mm
-        max_width = PAGE_WIDTH - right_margin - LEFT_MARGIN
-        if logo_width > max_width:
-            logo_height *= max_width / logo_width
-            logo_width = max_width
-        max_height = 19 * mm
-        if logo_height > max_height:
-            logo_width *= max_height / logo_height
-            logo_height = max_height
-        logo_alignment = logo_alignment.lower()
-        if logo_alignment == 'left':
-            x = LEFT_MARGIN
-        elif logo_alignment == 'center':
-            x = (LEFT_MARGIN + PAGE_WIDTH - right_margin - logo_width) / 2
-        else:
-            x = PAGE_WIDTH - right_margin - logo_width
-        y = PAGE_HEIGHT - 26.5 * mm
-        canvas.drawImage(logo_uri, x, y, logo_width, logo_height)
-    canvas.setFont('Helvetica', 8)
-    canvas.drawCentredString(PAGE_WIDTH - 17.5 * mm, PAGE_HEIGHT - 29 * mm, "#{}".format(object_id))
-    canvas.drawCentredString(PAGE_WIDTH - 17.5 * mm, PAGE_HEIGHT - 33 * mm, "Exported on")
-    canvas.drawCentredString(PAGE_WIDTH - 17.5 * mm, PAGE_HEIGHT - 37 * mm, datetime.datetime.utcnow().date().strftime('%Y-%m-%d'))
-    page_num = canvas.getPageNumber()
-    canvas.setFont('Helvetica', 11)
-    canvas.drawCentredString(PAGE_WIDTH / 2, 15 * mm, "{}".format(page_num))
-    canvas.restoreState()
-
-
-def _get_num_fitting_characters(canvas, line, max_width, font_name, font_size):
[…]
