+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Currently in development.
 - Allow linking to headers in Markdown content
 - Allow disabling of "Use in Measurement" button for samples
 - Added markdown support to object metadata
+- Reimplemented PDF export
 
 Version 0.16
 ------------
