### 2021-01-11 305dd98
Allow disabling of "Use in Measurement" button for samples
