+++ docs/administrator_guide/configuration.rst
@@ -173,5 +173,7 @@ Miscellaneous
      - If set, only administrators will be able to delete non-empty groups.
    * - SAMPLEDB_LOAD_OBJECTS_IN_BACKGROUND
      - If set, object selections will be loaded in the background using AJAX.
+   * - SAMPLEDB_DISABLE_USE_IN_MEASUREMENT
+     - If set, the "Use in Measurement" button will not be shown.
 
 There are other configuration values related to packages used by SampleDB. For more information on those, see the documentation of the corresponding packages.
