### 2021-01-12 055e75d
Add horizontal table style for arrays for objects
