### 2021-01-14 1e2daf7
Add a configuration variable to allow only administrators to create projects
