+++ sampledb/config.py
@@ -401,6 +401,7 @@ INVITATION_TIME_LIMIT = 7 * 24 * 60 * 60
 # other settings
 ONLY_ADMINS_CAN_MANAGE_LOCATIONS = False
 ONLY_ADMINS_CAN_DELETE_GROUPS = False
+ONLY_ADMINS_CAN_CREATE_PROJECTS = False
 
 DISABLE_USE_IN_MEASUREMENT = False
 
