+++ docs/administrator_guide/configuration.rst
@@ -171,6 +171,8 @@ Miscellaneous
      - If set, only administrators will be able to create and edit locations.
    * - SAMPLEDB_ONLY_ADMINS_CAN_DELETE_GROUPS
      - If set, only administrators will be able to delete non-empty groups.
+   * - SAMPLEDB_ONLY_ADMINS_CAN_CREATE_PROJECTS
+     - If set, only administrators will be able to create projects.
    * - SAMPLEDB_LOAD_OBJECTS_IN_BACKGROUND
      - If set, object selections will be loaded in the background using AJAX.
    * - SAMPLEDB_DISABLE_USE_IN_MEASUREMENT
