+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Currently in development.
 - Allow disabling of "Use in Measurement" button for samples
 - Added markdown support to object metadata
 - Reimplemented PDF export
+- Added a configuration variable to allow only administrators to create projects
 
 Version 0.16
 ------------
