### 2021-01-14 2e0ff2d
Add a configuration variable to allow only administrators to create groups
