+++ sampledb/config.py
@@ -400,6 +400,7 @@ INVITATION_TIME_LIMIT = 7 * 24 * 60 * 60
 
 # other settings
 ONLY_ADMINS_CAN_MANAGE_LOCATIONS = False
+ONLY_ADMINS_CAN_CREATE_GROUPS = False
 ONLY_ADMINS_CAN_DELETE_GROUPS = False
 ONLY_ADMINS_CAN_CREATE_PROJECTS = False
 
