+++ docs/administrator_guide/configuration.rst
@@ -169,6 +169,8 @@ Miscellaneous
      - The time limit for invitation links in seconds.
    * - SAMPLEDB_ONLY_ADMINS_CAN_MANAGE_LOCATIONS
      - If set, only administrators will be able to create and edit locations.
+   * - SAMPLEDB_ONLY_ADMINS_CAN_CREATE_GROUPS
+     - If set, only administrators will be able to create groups.
    * - SAMPLEDB_ONLY_ADMINS_CAN_DELETE_GROUPS
      - If set, only administrators will be able to delete non-empty groups.
    * - SAMPLEDB_ONLY_ADMINS_CAN_CREATE_PROJECTS
