### 2021-01-16 #2107 @arfon
:wave: @celliern @dvanic @schnorr, would any of you be willing to provide a second review for this submission to JOSS?

The title of the submission is _"SampleDB: A sample and measurement metadata database"_, which describes a [Python package](https://github.com/sciapp/sampledb) for managing experiments. Quoting from the SampleDB website:

> SampleDB is a web-based sample and measurement metadata database developed at PGI and JCNS. Scientists can use SampleDB to store and retrieve information on samples, measurements and simulations, analyze them using Jupyter notebooks, track sample locations and responsibilities and view sample lifecycles. As a tool for reproducible and sustainable science, it facilitates long-term storage of research data with clear benefits to scientists, interfaces with dedicated JupyterHub instances and can be integrated in instrument systems to automate data entry. A schema system allows users to store and search even complex types of metadata, and the powerful permission system ensures that information stays secure while simplifying collaboration with external users.

Our request here would be that you review the [software](https://github.com/sciapp/sampledb) and [short paper](https://github.com/openjournals/joss-papers/blob/joss.02107/joss.02107/10.21105.joss.02107.pdf), following the JOSS [checklist-driven review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Many thanks!
