### 2021-01-18 9580468
Improve handling of optional text input
