### 2021-01-20 c33583c
Improve error handling for admin only action types
