### 2021-01-20 689586b
Project permissions can be set when inviting a user
