### 2021-01-21 bc50519
Allow default value "self" for user fields
