### 2021-01-21 bed7105
Allow disabling of subprojects
