### 2021-01-26 94fea55
Rename projects to project groups and groups to basic groups to avoid ambiguity
