+++ sampledb/frontend/projects.py
@@ -22,7 +22,7 @@ def project(project_id):
         expiration_time_limit = flask.current_app.config['INVITATION_TIME_LIMIT']
         token_data = verify_token(token, salt='invite_to_project', secret_key=flask.current_app.config['SECRET_KEY'], expiration=expiration_time_limit)
         if token_data is None:
-            flask.flash('Invalid project invitation token. Please request a new invitation.', 'error')
+            flask.flash('Invalid project group invitation token. Please request a new invitation.', 'error')
             return flask.abort(403)
         if 'invitation_id' in token_data:
             if logic.projects.get_project_invitation(token_data['invitation_id']).accepted:
@@ -32,7 +32,7 @@ def project(project_id):
             return flask.abort(403)
         permissions = Permissions.from_value(token_data.get('permissions', Permissions.READ.value))
         if permissions == Permissions.NONE:
-            flask.flash('Invalid permissions in project invitation token. Please request a new invitation.', 'error')
+            flask.flash('Invalid permissions in project group invitation token. Please request a new invitation.', 'error')
             return flask.abort(403)
         user_id = token_data.get('user_id', None)
         if user_id != flask_login.current_user.id:
@@ -51,7 +51,7 @@ def project(project_id):
         try:
             logic.projects.add_user_to_project(project_id, user_id, permissions, other_project_ids=other_project_ids)
         except logic.errors.UserAlreadyMemberOfProjectError:
-            flask.flash('You are already a member of this project', 'error')
+            flask.flash('You are already a member of this project group.', 'error')
         except logic.errors.ProjectDoesNotExistError:
             pass
     user_id = flask_login.current_user.id
@@ -156,18 +156,18 @@ def project(project_id):
             try:
                 logic.projects.remove_user_from_project(project_id=project_id, user_id=user_id)
             except logic.errors.ProjectDoesNotExistError:
-                flask.flash('This project does not exist.', 'error')
+                flask.flash('This project group does not exist.', 'error')
                 return flask.redirect(flask.url_for('.projects'))
             except logic.errors.UserDoesNotExistError:
                 return flask.abort(500)
             except logic.errors.UserNotMemberOfProjectError:
-                flask.flash('You have already left the project.', 'error')
+                flask.flash('You have already left the project group.', 'error')
                 return flask.redirect(flask.url_for('.projects'))
             except logic.errors.NoMemberWithGrantPermissionsForProjectError:
-                flask.flash('You cannot leave this project, because your are the only user with GRANT permissions.', 'error')
+                flask.flash('You cannot leave this project group, because your are the only user with GRANT permissions.', 'error')
                 return flask.redirect(flask.url_for('.project', project_id=project_id))
             else:
-                flask.flash('You have successfully left the project.', 'success')
+                flask.flash('You have successfully left the project group.', 'success')
                 return flask.redirect(flask.url_for('.projects'))
     if 'delete' in flask.request.form and Permissions.GRANT in user_permissions:
         if delete_project_form.validate_on_submit():
@@ -175,10 +175,10 @@ def project(project_id):
             try:
                 logic.projects.delete_project(project_id=project_id)
             except logic.errors.ProjectDoesNotExistError:
-                flask.flash('This project has already been deleted.', 'success')
+                flask.flash('This project group has already been deleted.', 'success')
                 return flask.redirect(flask.url_for('.projects'))
             else:
-                flask.flash('You have successfully deleted the project.', 'success')
+                flask.flash('You have successfully deleted the project group.', 'success')
                 return flask.redirect(flask.url_for('.projects'))
     if 'remove_member' in flask.request.form and Permissions.GRANT in user_permissions:
[…]
