+++ sampledb/frontend/templates/mails/notifications/invited_to_group.txt
@@ -1,7 +1,7 @@
 {% extends "mails/notifications/base.txt" %}
 
 {% block message %}
-{{ get_user(data['inviter_id']).name }} (#{{ data['inviter_id'] }}) has invited you to be a member of the group "{{ get_group(data['group_id']).name}}". If you wish to join this group, please visit the following URL in your browser:
+{{ get_user(data['inviter_id']).name }} (#{{ data['inviter_id'] }}) has invited you to be a member of the basic group "{{ get_group(data['group_id']).name}}". If you wish to join this basic group, please visit the following URL in your browser:
 {{ data['confirmation_url'] }}
 
 This invitation will expire at {{ data['expiration_utc_datetime'] }}.

[…]
