### 2021-01-27 c760afe
Fix object name escaping when loading objects in the background
