### 2021-01-29 781640a
removed requirement to submit all possible keys to a quantity datatype object, reduces dependency on pint for external software importing data.
