### 2021-02-01 8e16841
Allow configuring the Help link
