### 2021-02-02 218916e
Add markdown support to instrument log
