### 2021-02-03 cf289a7
Fix action ID filtering when loading objects in the background
