### 2021-02-04 797fafc
* removed old commented code * dont override given instance values
