### 2021-02-04 175191a
* fixed stupid comparison * don't recalculate magnitude if creating object from json
