### 2021-02-05 bd7bb8c
Allow using data of referenced objects in notebook templates
