### 2021-02-10 1096b76
Fix action type filtering for loading in background
