### 2021-02-10 #2107 @FlorianRhiem
@arfon Thank you, I have created a new release (version 0.17.0), updated the paper.json accordingly and updated the auto-generated Zenodo entry to have the correct title and author: https://doi.org/10.5281/zenodo.4529206
