### 2021-02-10 #2107 @FlorianRhiem
The proof looks good to me. 👍 The top whedon post here still has version 0.11.0, but I didn't find that version referenced in either the paper or the crossref XML, so I guess that's fine and just for reviewing?
