### 2022-04-22 #2107 @eteq
While know this is like a year after the fact, it came up in a search for my recently and I just want to publicly apologies for ghosting y'all here.  This was definitely a victim of COVID/childcare/inability to keep up with committments in 2020, but I apologize all the same.
